@@ -2,7 +2,7 @@
 # future
 [metadata]
 name = fenics-ffcx
-version = 0.6.0
+version = 0.7.0
 author = FEniCS Project Contributors
 email = fenics-dev@googlegroups.com
 maintainer = FEniCS Project Steering Council
@@ -26,10 +26,10 @@ classifiers =
     Operating System :: MacOS :: MacOS X
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
     Topic :: Scientific/Engineering :: Mathematics
     Topic :: Software Development :: Libraries :: Python Modules
     Topic :: Software Development :: Code Generators
@@ -38,16 +38,16 @@ classifiers =
 packages = find:
 include_package_data = True
 zip_safe = False
-python_requires = >= 3.7
+python_requires = >= 3.8
 setup_requires =
-    setuptools >= 58
+    setuptools >= 62
     wheel
 install_requires =
     numpy
     cffi
     setuptools
-    fenics-basix >= 0.6.0.dev0, <0.7.0
-    fenics-ufl >= 2022.3.0.dev0, <2023.1.1
+    fenics-basix >= 0.7.0, <0.8.0
+    fenics-ufl >= 2023.2.0, <2023.3.0
 
 [options.extras_require]
 docs = sphinx; sphinx_rtd_theme
