@@ -19,18 +19,19 @@
 # a set of interpolation points
 
 import basix
-from ufl import (Coefficient, FiniteElement, FunctionSpace, Mesh, MixedElement,
-                 VectorElement, grad, triangle)
+import basix.ufl
+from ffcx.element_interface import QuadratureElement
+from ufl import Coefficient, FunctionSpace, Mesh, grad
 
 # Define mesh
-cell = triangle
-v_el = VectorElement("Lagrange", cell, 1)
+cell = "triangle"
+v_el = basix.ufl.element("Lagrange", cell, 1, shape=(2, ))
 mesh = Mesh(v_el)
 
 # Define mixed function space
-el = FiniteElement("CG", cell, 2)
-el_int = VectorElement("Discontinuous Lagrange", cell, 1)
-me = MixedElement([el, el_int])
+el = basix.ufl.element("P", cell, 2)
+el_int = basix.ufl.element("Discontinuous Lagrange", cell, 1, shape=(2, ))
+me = basix.ufl.mixed_element([el, el_int])
 V = FunctionSpace(mesh, me)
 u = Coefficient(V)
 
@@ -41,21 +42,21 @@ du1 = grad(u[1])
 # Define an expression using quadrature elements
 q_rule = "gauss_jacobi"
 q_degree = 3
-q_el = FiniteElement("Quadrature", cell, q_degree, quad_scheme=q_rule)
+q_el = QuadratureElement(cell, (), q_rule, q_degree)
 Q = FunctionSpace(mesh, q_el)
 q = Coefficient(Q)
 powq = 3 * q**2
 
 # Extract basix cell type
-b_cell = basix.cell.string_to_type(cell.cellname())
+b_cell = basix.cell.string_to_type(cell)
 
 # Find quadrature points for quadrature element
 b_rule = basix.quadrature.string_to_type(q_rule)
-quadrature_points, _ = basix.make_quadrature(b_rule, b_cell, q_degree)
+quadrature_points, _ = basix.quadrature.make_quadrature(b_cell, q_degree, rule=b_rule)
 
 # Get interpolation points for output space
-family = basix.finite_element.string_to_family("Lagrange", cell.cellname())
-b_element = basix.create_element(family, b_cell, 4, basix.LagrangeVariant.gll_warped, True)
+family = basix.finite_element.string_to_family("Lagrange", cell)
+b_element = basix.create_element(family, b_cell, 4, basix.LagrangeVariant.gll_warped, discontinuous=True)
 interpolation_points = b_element.points
 
 # Create expressions that can be used for interpolation
