@@ -16,11 +16,11 @@
 # along with FFCx. If not, see <http://www.gnu.org/licenses/>.
 #
 # Test form for metadata.
-from ufl import (Coefficient, FiniteElement, TestFunction, TrialFunction,
-                 VectorElement, dx, grad, inner, triangle)
+import basix.ufl
+from ufl import Coefficient, TestFunction, TrialFunction, dx, grad, inner
 
-element = FiniteElement("Lagrange", triangle, 1)
-vector_element = VectorElement("Lagrange", triangle, 1)
+element = basix.ufl.element("Lagrange", "triangle", 1)
+vector_element = basix.ufl.element("Lagrange", "triangle", 1, shape=(2, ))
 
 
 u = TrialFunction(element)
@@ -32,3 +32,5 @@ a = inner(grad(u), grad(v)) * dx(0, degree=8)\
     + inner(c, c) * inner(grad(u), grad(v)) * dx(1, degree=4)\
     + inner(c, c) * inner(grad(u), grad(v)) * dx(1, degree=2)\
     + inner(grad(u), grad(v)) * dx(1, degree=-1)
+
+L = v * dx(0, metadata={"precision": 1})
