@@ -18,16 +18,14 @@
 # along with FFCx. If not, see <http://www.gnu.org/licenses/>.
 #
 # Mixed coefficient.
-from ufl import (Coefficients, FiniteElement, MixedElement, VectorElement, dot,
-                 dS, dx, triangle)
+import basix.ufl
+from ufl import Coefficients, dot, dS, dx
 
-cell = triangle
+DG = basix.ufl.element("DG", "triangle", 0, shape=(2, ))
+CG = basix.ufl.element("Lagrange", "triangle", 2)
+RT = basix.ufl.element("RT", "triangle", 3)
 
-DG = VectorElement("DG", cell, 0)
-CG = FiniteElement("Lagrange", cell, 2)
-RT = FiniteElement("RT", cell, 3)
-
-element = MixedElement(DG, CG, RT)
+element = basix.ufl.mixed_element([DG, CG, RT])
 
 f, g, h = Coefficients(element)
 
