@@ -17,10 +17,10 @@
 #
 # The bilinear form a(u, v) and linear form L(v) for a simple
 # reaction-diffusion equation using simplified tuple notation.
-from ufl import (Coefficient, FiniteElement, TestFunction, TrialFunction, dx,
-                 grad, inner, triangle)
+import basix.ufl
+from ufl import Coefficient, TestFunction, TrialFunction, dx, grad, inner
 
-element = FiniteElement("Lagrange", triangle, 1)
+element = basix.ufl.element("Lagrange", "triangle", 1)
 
 u = TrialFunction(element)
 v = TestFunction(element)
