@@ -10,8 +10,8 @@
 
 import logging
 
-import ffcx.codegeneration.basix_custom_element_template as ufcx_basix_custom_finite_element
-import ffcx.codegeneration.finite_element_template as ufcx_finite_element
+import ffcx.codegeneration.C.basix_custom_element_template as ufcx_basix_custom_finite_element
+import ffcx.codegeneration.C.finite_element_template as ufcx_finite_element
 import ufl
 
 logger = logging.getLogger("ffcx")
@@ -63,30 +63,29 @@ def generator(ir, options):
     else:
         d["basix_cell"] = int(ir.basix_cell)
 
-    import ffcx.codegeneration.C.cnodes as L
-
     if len(ir.value_shape) > 0:
         d["value_shape"] = f"value_shape_{ir.name}"
-        d["value_shape_init"] = L.ArrayDecl(
-            "int", f"value_shape_{ir.name}", values=ir.value_shape, sizes=len(ir.value_shape))
+        values = ", ".join(str(i) for i in ir.value_shape)
+        sizes = len(ir.value_shape)
+        d["value_shape_init"] = f"int value_shape_{ir.name}[{sizes}] = {{{values}}};"
     else:
         d["value_shape"] = "NULL"
         d["value_shape_init"] = ""
 
     if len(ir.reference_value_shape) > 0:
         d["reference_value_shape"] = f"reference_value_shape_{ir.name}"
-        d["reference_value_shape_init"] = L.ArrayDecl(
-            "int", f"reference_value_shape_{ir.name}",
-            values=ir.reference_value_shape, sizes=len(ir.reference_value_shape))
+        values = ", ".join(str(i) for i in ir.reference_value_shape)
+        sizes = len(ir.reference_value_shape)
+        d["reference_value_shape_init"] = f"int reference_value_shape_{ir.name}[{sizes}] = {{{values}}};"
     else:
         d["reference_value_shape"] = "NULL"
         d["reference_value_shape_init"] = ""
 
     if len(ir.sub_elements) > 0:
         d["sub_elements"] = f"sub_elements_{ir.name}"
-        d["sub_elements_init"] = L.ArrayDecl(
-            "ufcx_finite_element*", f"sub_elements_{ir.name}",
-            values=[L.AddressOf(L.Symbol(el)) for el in ir.sub_elements], sizes=len(ir.sub_elements))
+        values = ", ".join(f"&{el}" for el in ir.sub_elements)
+        sizes = len(ir.sub_elements)
+        d["sub_elements_init"] = f"ufcx_finite_element* sub_elements_{ir.name}[{sizes}] = {{{values}}};"
     else:
         d["sub_elements"] = "NULL"
         d["sub_elements_init"] = ""
@@ -119,20 +118,19 @@ def generate_custom_element(name, ir):
     d = {}
     d["factory_name"] = name
     d["cell_type"] = int(ir.cell_type)
+    d["polyset_type"] = int(ir.polyset_type)
     d["map_type"] = int(ir.map_type)
     d["sobolev_space"] = int(ir.sobolev_space)
     d["highest_complete_degree"] = ir.highest_complete_degree
     d["highest_degree"] = ir.highest_degree
     d["discontinuous"] = "true" if ir.discontinuous else "false"
     d["interpolation_nderivs"] = ir.interpolation_nderivs
-
-    import ffcx.codegeneration.C.cnodes as L
-
     d["value_shape_length"] = len(ir.value_shape)
     if len(ir.value_shape) > 0:
         d["value_shape"] = f"value_shape_{name}"
-        d["value_shape_init"] = L.ArrayDecl(
-            "int", f"value_shape_{name}", values=ir.value_shape, sizes=len(ir.value_shape))
+        values = ", ".join(str(i) for i in ir.value_shape)
+        sizes = len(ir.value_shape)
+        d["value_shape_init"] = f"int value_shape_{name}[{sizes}] = {{{values}}};"
     else:
         d["value_shape"] = "NULL"
         d["value_shape_init"] = ""
