@@ -24,6 +24,9 @@ factory = """
 {original_coefficient_position_init}
 {dofmaps_init}
 {finite_elements_init}
+{form_integral_offsets_init}
+{form_integrals_init}
+{form_integral_ids_init}
 
 // Return a list of the coefficient names.
 const char** coefficient_name_{factory_name}(void)
@@ -35,21 +38,6 @@ const char** coefficient_name_{factory_name}(void)
 const char** constant_name_{factory_name}(void)
 {{
 {constant_name_map}
-}}
-
-int* integral_ids_{factory_name}(ufcx_integral_type integral_type)
-{{
-{integral_ids}
-}}
-
-int num_integrals_{factory_name}(ufcx_integral_type integral_type)
-{{
-{num_integrals}
-}}
-
-ufcx_integral** integrals_{factory_name}(ufcx_integral_type integral_type)
-{{
-{integrals}
 }}
 
 ufcx_form {factory_name} =
@@ -67,10 +55,9 @@ ufcx_form {factory_name} =
   .finite_elements = {finite_elements},
   .dofmaps = {dofmaps},
 
-  .integral_ids = integral_ids_{factory_name},
-  .num_integrals = num_integrals_{factory_name},
-
-  .integrals = integrals_{factory_name}
+  .form_integrals = {form_integrals},
+  .form_integral_ids = {form_integral_ids},
+  .form_integral_offsets = form_integral_offsets_{factory_name}
 }};
 
 // Alias name
