@@ -5,11 +5,7 @@
 # SPDX-License-Identifier:    LGPL-3.0-or-later
 """Collection of FFCx specific pieces for the code generation phase."""
 
-import types
-
-import ffcx.codegeneration.C.cnodes
 from ffcx.codegeneration.access import FFCXBackendAccess
-from ffcx.codegeneration.C.ufl_to_cnodes import UFL2CNodesTranslatorCpp
 from ffcx.codegeneration.definitions import FFCXBackendDefinitions
 from ffcx.codegeneration.symbols import FFCXBackendSymbols
 
@@ -19,19 +15,12 @@ class FFCXBackend(object):
 
     def __init__(self, ir, options):
 
-        # This is the seam where cnodes/C is chosen for the FFCx backend
-        self.language: types.ModuleType = ffcx.codegeneration.C.cnodes
-        scalar_type = options["scalar_type"]
-        self.ufl_to_language = UFL2CNodesTranslatorCpp(self.language, scalar_type)
-
         coefficient_numbering = ir.coefficient_numbering
         coefficient_offsets = ir.coefficient_offsets
 
         original_constant_offsets = ir.original_constant_offsets
 
-        self.symbols = FFCXBackendSymbols(self.language, coefficient_numbering,
+        self.symbols = FFCXBackendSymbols(coefficient_numbering,
                                           coefficient_offsets, original_constant_offsets)
-        self.definitions = FFCXBackendDefinitions(ir, self.language,
-                                                  self.symbols, options)
-        self.access = FFCXBackendAccess(ir, self.language, self.symbols,
-                                        options)
+        self.definitions = FFCXBackendDefinitions(ir, self.symbols, options)
+        self.access = FFCXBackendAccess(ir, self.symbols, options)
