@@ -14,12 +14,13 @@ UFC function from an intermediate representation (IR).
 import logging
 import typing
 
-from ffcx.codegeneration.dofmap import generator as dofmap_generator
-from ffcx.codegeneration.expressions import generator as expression_generator
-from ffcx.codegeneration.finite_element import \
+from ffcx.codegeneration.C.dofmap import generator as dofmap_generator
+from ffcx.codegeneration.C.expressions import generator as expression_generator
+from ffcx.codegeneration.C.finite_element import \
     generator as finite_element_generator
-from ffcx.codegeneration.form import generator as form_generator
-from ffcx.codegeneration.integrals import generator as integral_generator
+from ffcx.codegeneration.C.form import generator as form_generator
+from ffcx.codegeneration.C.integrals import generator as integral_generator
+from ffcx.codegeneration.C.file import generator as file_generator
 
 logger = logging.getLogger("ffcx")
 
@@ -28,14 +29,17 @@ class CodeBlocks(typing.NamedTuple):
     """
     Storage of code blocks of the form (declaration, implementation).
 
-    Blocks for elements, dofmaps, integrals, forms and expressions is stored
+    Blocks for elements, dofmaps, integrals, forms and expressions,
+    and start and end of file output
     """
 
+    file_pre: typing.List[typing.Tuple[str, str]]
     elements: typing.List[typing.Tuple[str, str]]
     dofmaps: typing.List[typing.Tuple[str, str]]
     integrals: typing.List[typing.Tuple[str, str]]
     forms: typing.List[typing.Tuple[str, str]]
     expressions: typing.List[typing.Tuple[str, str]]
+    file_post: typing.List[typing.Tuple[str, str]]
 
 
 def generate_code(ir, options) -> CodeBlocks:
@@ -50,5 +54,7 @@ def generate_code(ir, options) -> CodeBlocks:
     code_integrals = [integral_generator(integral_ir, options) for integral_ir in ir.integrals]
     code_forms = [form_generator(form_ir, options) for form_ir in ir.forms]
     code_expressions = [expression_generator(expression_ir, options) for expression_ir in ir.expressions]
-    return CodeBlocks(elements=code_finite_elements, dofmaps=code_dofmaps,
-                      integrals=code_integrals, forms=code_forms, expressions=code_expressions)
+    code_file_pre, code_file_post = file_generator(options)
+    return CodeBlocks(file_pre=[code_file_pre], elements=code_finite_elements, dofmaps=code_dofmaps,
+                      integrals=code_integrals, forms=code_forms, expressions=code_expressions,
+                      file_post=[code_file_post])
