@@ -8,8 +8,8 @@
 import logging
 
 import ufl
-from ffcx.element_interface import create_element
-from ffcx.naming import scalar_to_value_type
+from ffcx.element_interface import convert_element
+import ffcx.codegeneration.lnodes as L
 
 logger = logging.getLogger("ffcx")
 
@@ -17,11 +17,10 @@ logger = logging.getLogger("ffcx")
 class FFCXBackendDefinitions(object):
     """FFCx specific code definitions."""
 
-    def __init__(self, ir, language, symbols, options):
+    def __init__(self, ir, symbols, options):
         # Store ir and options
         self.integral_type = ir.integral_type
         self.entitytype = ir.entitytype
-        self.language = language
         self.symbols = symbols
         self.options = options
 
@@ -64,8 +63,6 @@ class FFCXBackendDefinitions(object):
 
     def coefficient(self, t, mt, tabledata, quadrature_rule, access):
         """Return definition code for coefficients."""
-        L = self.language
-
         ttype = tabledata.ttype
         num_dofs = tabledata.values.shape[3]
         bs = tabledata.block_size
@@ -84,7 +81,7 @@ class FFCXBackendDefinitions(object):
 
         # Get access to element table
         FE = self.symbols.element_table(tabledata, self.entitytype, mt.restriction)
-        ic = self.symbols.coefficient_dof_sum_index()
+        ic = self.symbols.coefficient_dof_sum_index
 
         code = []
         pre_code = []
@@ -99,14 +96,14 @@ class FFCXBackendDefinitions(object):
 
             # If a map is necessary from stride 1 to bs, the code must be added before the quadrature loop.
             if dof_access_map:
-                pre_code += [L.ArrayDecl(self.options["scalar_type"], dof_access.array, num_dofs)]
-                pre_body = L.Assign(dof_access, dof_access_map)
+                pre_code += [L.ArrayDecl(dof_access.array, sizes=num_dofs)]
+                pre_body = [L.Assign(dof_access, dof_access_map)]
                 pre_code += [L.ForRange(ic, 0, num_dofs, pre_body)]
         else:
             dof_access = self.symbols.coefficient_dof_access(mt.terminal, ic * bs + begin)
 
         body = [L.AssignAdd(access, dof_access * FE[ic])]
-        code += [L.VariableDecl(self.options["scalar_type"], access, 0.0)]
+        code += [L.VariableDecl(access, 0.0)]
         code += [L.ForRange(ic, 0, num_dofs, body)]
 
         return pre_code, code
@@ -119,12 +116,10 @@ class FFCXBackendDefinitions(object):
 
     def _define_coordinate_dofs_lincomb(self, e, mt, tabledata, quadrature_rule, access):
         """Define x or J as a linear combination of coordinate dofs with given table data."""
-        L = self.language
-
         # Get properties of domain
         domain = ufl.domain.extract_unique_domain(mt.terminal)
         coordinate_element = domain.ufl_coordinate_element()
-        num_scalar_dofs = create_element(coordinate_element).sub_element.dim
+        num_scalar_dofs = convert_element(coordinate_element).sub_element.dim
 
         num_dofs = tabledata.values.shape[3]
         begin = tabledata.offset
@@ -139,20 +134,12 @@ class FFCXBackendDefinitions(object):
 
         # Get access to element table
         FE = self.symbols.element_table(tabledata, self.entitytype, mt.restriction)
-        ic = self.symbols.coefficient_dof_sum_index()
-        dof_access = self.symbols.S("coordinate_dofs")
-
-        # coordinate dofs is always 3d
-        dim = 3
-        offset = 0
-        if mt.restriction == "-":
-            offset = num_scalar_dofs * dim
-
-        value_type = scalar_to_value_type(self.options["scalar_type"])
+        ic = self.symbols.coefficient_dof_sum_index
+        dof_access = self.symbols.domain_dof_access(ic, begin, 3, num_scalar_dofs, mt.restriction)
 
         code = []
-        body = [L.AssignAdd(access, dof_access[ic * dim + begin + offset] * FE[ic])]
-        code += [L.VariableDecl(f"{value_type}", access, 0.0)]
+        body = [L.AssignAdd(access, dof_access * FE[ic])]
+        code += [L.VariableDecl(access, 0.0)]
         code += [L.ForRange(ic, 0, num_scalar_dofs, body)]
 
         return [], code
