@@ -10,126 +10,13 @@ from itertools import product
 from typing import Any, DefaultDict, Dict, Set
 
 import ufl
-from ffcx.codegeneration import expressions_template, geometry
+from ffcx.codegeneration import geometry
 from ffcx.codegeneration.backend import FFCXBackend
-from ffcx.codegeneration.C.cnodes import CNode
-from ffcx.codegeneration.C.format_lines import format_indented_lines
+import ffcx.codegeneration.lnodes as L
+from ffcx.codegeneration.lnodes import LNode
 from ffcx.ir.representation import ExpressionIR
-from ffcx.naming import cdtype_to_numpy, scalar_to_value_type
 
 logger = logging.getLogger("ffcx")
-
-
-def generator(ir, options):
-    """Generate UFC code for an expression."""
-    logger.info("Generating code for expression:")
-    logger.info(f"--- points: {ir.points}")
-    logger.info(f"--- name: {ir.name}")
-
-    factory_name = ir.name
-
-    # Format declaration
-    declaration = expressions_template.declaration.format(
-        factory_name=factory_name, name_from_uflfile=ir.name_from_uflfile)
-
-    backend = FFCXBackend(ir, options)
-    L = backend.language
-    eg = ExpressionGenerator(ir, backend)
-
-    d = {}
-    d["name_from_uflfile"] = ir.name_from_uflfile
-    d["factory_name"] = ir.name
-
-    parts = eg.generate()
-
-    body = format_indented_lines(parts.cs_format(), 1)
-    d["tabulate_expression"] = body
-
-    if len(ir.original_coefficient_positions) > 0:
-        d["original_coefficient_positions"] = f"original_coefficient_positions_{ir.name}"
-        d["original_coefficient_positions_init"] = L.ArrayDecl(
-            "static int", f"original_coefficient_positions_{ir.name}",
-            values=ir.original_coefficient_positions, sizes=len(ir.original_coefficient_positions))
-    else:
-        d["original_coefficient_positions"] = L.Null()
-        d["original_coefficient_positions_init"] = ""
-
-    d["points_init"] = L.ArrayDecl(
-        "static double", f"points_{ir.name}", values=ir.points.flatten(), sizes=ir.points.size)
-    d["points"] = L.Symbol(f"points_{ir.name}")
-
-    if len(ir.expression_shape) > 0:
-        d["value_shape_init"] = L.ArrayDecl(
-            "static int", f"value_shape_{ir.name}", values=ir.expression_shape, sizes=len(ir.expression_shape))
-        d["value_shape"] = f"value_shape_{ir.name}"
-    else:
-        d["value_shape_init"] = ""
-        d["value_shape"] = L.Null()
-
-    d["num_components"] = len(ir.expression_shape)
-    d["num_coefficients"] = len(ir.coefficient_numbering)
-    d["num_constants"] = len(ir.constant_names)
-    d["num_points"] = ir.points.shape[0]
-    d["topological_dimension"] = ir.points.shape[1]
-    d["scalar_type"] = options["scalar_type"]
-    d["geom_type"] = scalar_to_value_type(options["scalar_type"])
-    d["np_scalar_type"] = cdtype_to_numpy(options["scalar_type"])
-
-    d["rank"] = len(ir.tensor_shape)
-
-    if len(ir.coefficient_names) > 0:
-        d["coefficient_names_init"] = L.ArrayDecl(
-            "static const char*", f"coefficient_names_{ir.name}", values=ir.coefficient_names,
-            sizes=len(ir.coefficient_names))
-        d["coefficient_names"] = f"coefficient_names_{ir.name}"
-    else:
-        d["coefficient_names_init"] = ""
-        d["coefficient_names"] = L.Null()
-
-    if len(ir.constant_names) > 0:
-        d["constant_names_init"] = L.ArrayDecl(
-            "static const char*", f"constant_names_{ir.name}", values=ir.constant_names,
-            sizes=len(ir.constant_names))
-        d["constant_names"] = f"constant_names_{ir.name}"
-    else:
-        d["constant_names_init"] = ""
-        d["constant_names"] = L.Null()
-
-    code = []
-
-    # FIXME: Should be handled differently, revise how
-    # ufcx_function_space is generated (also for ufcx_form)
-    for (name, (element, dofmap, cmap_family, cmap_degree)) in ir.function_spaces.items():
-        code += [f"static ufcx_function_space function_space_{name}_{ir.name_from_uflfile} ="]
-        code += ["{"]
-        code += [f".finite_element = &{element},"]
-        code += [f".dofmap = &{dofmap},"]
-        code += [f".geometry_family = \"{cmap_family}\","]
-        code += [f".geometry_degree = {cmap_degree}"]
-        code += ["};"]
-
-    d["function_spaces_alloc"] = L.StatementList(code)
-    d["function_spaces"] = ""
-
-    if len(ir.function_spaces) > 0:
-        d["function_spaces"] = f"function_spaces_{ir.name}"
-        d["function_spaces_init"] = L.ArrayDecl("ufcx_function_space*", f"function_spaces_{ir.name}", values=[
-                                                L.AddressOf(L.Symbol(f"function_space_{name}_{ir.name_from_uflfile}"))
-                                                for (name, _) in ir.function_spaces.items()],
-                                                sizes=len(ir.function_spaces))
-    else:
-        d["function_spaces"] = L.Null()
-        d["function_spaces_init"] = ""
-
-    # Check that no keys are redundant or have been missed
-    from string import Formatter
-    fields = [fname for _, fname, _, _ in Formatter().parse(expressions_template.factory) if fname]
-    assert set(fields) == set(d.keys()), "Mismatch between keys in template and in formatting dict"
-
-    # Format implementation code
-    implementation = expressions_template.factory.format_map(d)
-
-    return declaration, implementation
 
 
 class ExpressionGenerator:
@@ -140,22 +27,17 @@ class ExpressionGenerator:
 
         self.ir = ir
         self.backend = backend
-        self.scope: Dict[Any, CNode] = {}
+        self.scope: Dict[Any, LNode] = {}
         self._ufl_names: Set[Any] = set()
         self.symbol_counters: DefaultDict[Any, int] = collections.defaultdict(int)
         self.shared_symbols: Dict[Any, Any] = {}
         self.quadrature_rule = list(self.ir.integrand.keys())[0]
 
     def generate(self):
-        L = self.backend.language
-
         parts = []
-        scalar_type = self.backend.access.options["scalar_type"]
-        value_type = scalar_to_value_type(scalar_type)
-
-        parts += self.generate_element_tables(value_type)
+        parts += self.generate_element_tables()
         # Generate the tables of geometry data that are needed
-        parts += self.generate_geometry_tables(value_type)
+        parts += self.generate_geometry_tables()
         parts += self.generate_piecewise_partition()
 
         all_preparts = []
@@ -171,10 +53,8 @@ class ExpressionGenerator:
 
         return L.StatementList(parts)
 
-    def generate_geometry_tables(self, float_type: str):
+    def generate_geometry_tables(self):
         """Generate static tables of geometry data."""
-        L = self.backend.language
-
         # Currently we only support circumradius
         ufl_geometry = {
             ufl.geometry.ReferenceCellVolume: "reference_cell_volume",
@@ -192,24 +72,22 @@ class ExpressionGenerator:
         parts = []
         for i, cell_list in cells.items():
             for c in cell_list:
-                parts.append(geometry.write_table(L, ufl_geometry[i], c, float_type))
+                parts.append(geometry.write_table(ufl_geometry[i], c))
 
         return parts
 
-    def generate_element_tables(self, float_type: str):
+    def generate_element_tables(self):
         """Generate tables of FE basis evaluated at specified points."""
-        L = self.backend.language
         parts = []
 
         tables = self.ir.unique_tables
-
-        padlen = self.ir.options["padlen"]
         table_names = sorted(tables)
 
         for name in table_names:
             table = tables[name]
-            decl = L.ArrayDecl(
-                f"static const {float_type}", name, table.shape, table, padlen=padlen)
+            symbol = L.Symbol(name, dtype=L.DataType.REAL)
+            self.backend.symbols.element_tables[name] = symbol
+            decl = L.ArrayDecl(symbol, sizes=table.shape, values=table, const=True)
             parts += [decl]
 
         # Add leading comment if there are any tables
@@ -225,8 +103,6 @@ class ExpressionGenerator:
         In the context of expressions quadrature loop is not accumulated.
 
         """
-        L = self.backend.language
-
         # Generate varying partition
         body = self.generate_varying_partition()
         body = L.commented_code_list(
@@ -243,7 +119,7 @@ class ExpressionGenerator:
             # Could happen for integral with everything zero and optimized away
             quadparts = []
         else:
-            iq = self.backend.symbols.quadrature_loop_index()
+            iq = self.backend.symbols.quadrature_loop_index
             num_points = self.quadrature_rule.points.shape[0]
             quadparts = [L.ForRange(iq, 0, num_points, body=body)]
 
@@ -251,12 +127,10 @@ class ExpressionGenerator:
 
     def generate_varying_partition(self):
         """Generate factors of blocks which are not cellwise constant."""
-        L = self.backend.language
-
         # Get annotated graph of factorisation
         F = self.ir.integrand[self.quadrature_rule]["factorization"]
 
-        arraysymbol = L.Symbol(f"sv_{self.quadrature_rule.id()}")
+        arraysymbol = L.Symbol(f"sv_{self.quadrature_rule.id()}", dtype=L.DataType.SCALAR)
         parts = self.generate_partition(arraysymbol, F, "varying")
         parts = L.commented_code_list(
             parts, f"Unstructured varying computations for quadrature rule {self.quadrature_rule.id()}")
@@ -264,12 +138,10 @@ class ExpressionGenerator:
 
     def generate_piecewise_partition(self):
         """Generate factors of blocks which are constant (i.e. do not depend on quadrature points)."""
-        L = self.backend.language
-
         # Get annotated graph of factorisation
         F = self.ir.integrand[self.quadrature_rule]["factorization"]
 
-        arraysymbol = L.Symbol("sp")
+        arraysymbol = L.Symbol("sp", dtype=L.DataType.SCALAR)
         parts = self.generate_partition(arraysymbol, F, "piecewise")
         parts = L.commented_code_list(parts, "Unstructured piecewise computations")
         return parts
@@ -301,8 +173,6 @@ class ExpressionGenerator:
 
     def generate_block_parts(self, blockmap, blockdata):
         """Generate and return code parts for a given block."""
-        L = self.backend.language
-
         # The parts to return
         preparts = []
         quadparts = []
@@ -322,11 +192,9 @@ class ExpressionGenerator:
         components = ufl.product(self.ir.expression_shape)
 
         num_points = self.quadrature_rule.points.shape[0]
-        A_shape = self.ir.tensor_shape
-        Asym = self.backend.symbols.element_tensor()
-        A = L.FlattenedArray(Asym, dims=[num_points, components] + A_shape)
-
-        iq = self.backend.symbols.quadrature_loop_index()
+        A_shape = [num_points, components] + self.ir.tensor_shape
+        A = self.backend.symbols.element_tensor
+        iq = self.backend.symbols.quadrature_loop_index
 
         # Check if DOFs in dofrange are equally spaced.
         expand_loop = False
@@ -349,7 +217,8 @@ class ExpressionGenerator:
                     f = self.get_var(F.nodes[fi_ci[0]]["expression"])
                     arg_factors = self.get_arg_factors(blockdata, block_rank, B_indices)
                     Brhs = L.float_product([f] + arg_factors)
-                    quadparts.append(L.AssignAdd(A[(A_indices[0], fi_ci[1]) + A_indices[1:]], Brhs))
+                    multi_index = L.MultiIndex([A_indices[0], fi_ci[1]] + A_indices[1:], A_shape)
+                    quadparts.append(L.AssignAdd(A[multi_index], Brhs))
         else:
 
             # Prepend dimensions of dofmap block with free index
@@ -381,7 +250,9 @@ class ExpressionGenerator:
             for fi_ci in blockdata.factor_indices_comp_indices:
                 f = self.get_var(F.nodes[fi_ci[0]]["expression"])
                 Brhs = L.float_product([f] + arg_factors)
-                body.append(L.AssignAdd(A[(A_indices[0], fi_ci[1]) + A_indices[1:]], Brhs))
+                indices = [A_indices[0], fi_ci[1]] + list(A_indices[1:])
+                multi_index = L.MultiIndex(indices, A_shape)
+                body.append(L.AssignAdd(A[multi_index], Brhs))
 
             for i in reversed(range(block_rank)):
                 body = L.ForRange(
@@ -401,8 +272,6 @@ class ExpressionGenerator:
             Indices used to index element tables
 
         """
-        L = self.backend.language
-
         arg_factors = []
         for i in range(block_rank):
             mad = blockdata.ma_data[i]
@@ -422,21 +291,18 @@ class ExpressionGenerator:
 
     def new_temp_symbol(self, basename):
         """Create a new code symbol named basename + running counter."""
-        L = self.backend.language
         name = "%s%d" % (basename, self.symbol_counters[basename])
         self.symbol_counters[basename] += 1
-        return L.Symbol(name)
+        return L.Symbol(name, dtype=L.DataType.SCALAR)
 
     def get_var(self, v):
         if v._ufl_is_literal_:
-            return self.backend.ufl_to_language.get(v)
+            return L.ufl_to_lnodes(v)
         f = self.scope.get(v)
         return f
 
     def generate_partition(self, symbol, F, mode):
         """Generate computations of factors of blocks."""
-        L = self.backend.language
-
         definitions = []
         pre_definitions = dict()
         intermediates = []
@@ -450,7 +316,7 @@ class ExpressionGenerator:
             mt = attr.get('mt')
 
             if v._ufl_is_literal_:
-                vaccess = self.backend.ufl_to_language.get(v)
+                vaccess = L.ufl_to_lnodes(v)
             elif mt is not None:
                 # All finite element based terminals have table data, as well
                 # as some, but not all, of the symbolic geometric terminals
@@ -479,7 +345,7 @@ class ExpressionGenerator:
 
                 # Mapping UFL operator to target language
                 self._ufl_names.add(v._ufl_handler_name_)
-                vexpr = self.backend.ufl_to_language.get(v, *vops)
+                vexpr = L.ufl_to_lnodes(v, *vops)
 
                 # Create a new intermediate for each subexpression
                 # except boolean conditions and its childs
@@ -505,7 +371,7 @@ class ExpressionGenerator:
                         intermediates.append(L.Assign(vaccess, vexpr))
                     else:
                         scalar_type = self.backend.access.options["scalar_type"]
-                        vaccess = L.Symbol("%s_%d" % (symbol.name, j))
+                        vaccess = L.Symbol("%s_%d" % (symbol.name, j), dtype=L.DataType.SCALAR)
                         intermediates.append(L.VariableDecl(f"const {scalar_type}", vaccess, vexpr))
 
             # Store access node for future reference
@@ -523,7 +389,6 @@ class ExpressionGenerator:
 
         if intermediates:
             if use_symbol_array:
-                scalar_type = self.backend.access.options["scalar_type"]
-                parts += [L.ArrayDecl(scalar_type, symbol, len(intermediates))]
+                parts += [L.ArrayDecl(symbol, sizes=len(intermediates))]
             parts += intermediates
         return parts
