@@ -4,32 +4,32 @@
 #
 # SPDX-License-Identifier:    LGPL-3.0-or-later
 
-import numpy
-
+import numpy as np
+import ffcx.codegeneration.lnodes as L
 import basix
 
 
-def write_table(L, tablename, cellname, type: str):
+def write_table(tablename, cellname):
     if tablename == "facet_edge_vertices":
-        return facet_edge_vertices(L, tablename, cellname)
+        return facet_edge_vertices(tablename, cellname)
     if tablename == "reference_facet_jacobian":
-        return reference_facet_jacobian(L, tablename, cellname, type)
+        return reference_facet_jacobian(tablename, cellname)
     if tablename == "reference_cell_volume":
-        return reference_cell_volume(L, tablename, cellname, type)
+        return reference_cell_volume(tablename, cellname)
     if tablename == "reference_facet_volume":
-        return reference_facet_volume(L, tablename, cellname, type)
+        return reference_facet_volume(tablename, cellname)
     if tablename == "reference_edge_vectors":
-        return reference_edge_vectors(L, tablename, cellname, type)
+        return reference_edge_vectors(tablename, cellname)
     if tablename == "facet_reference_edge_vectors":
-        return facet_reference_edge_vectors(L, tablename, cellname, type)
+        return facet_reference_edge_vectors(tablename, cellname)
     if tablename == "reference_facet_normals":
-        return reference_facet_normals(L, tablename, cellname, type)
+        return reference_facet_normals(tablename, cellname)
     if tablename == "facet_orientation":
-        return facet_orientation(L, tablename, cellname, type)
+        return facet_orientation(tablename, cellname)
     raise ValueError(f"Unknown geometry table name: {tablename}")
 
 
-def facet_edge_vertices(L, tablename, cellname):
+def facet_edge_vertices(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     topology = basix.topology(celltype)
     triangle_edges = basix.topology(basix.CellType.triangle)[1]
@@ -41,47 +41,52 @@ def facet_edge_vertices(L, tablename, cellname):
     edge_vertices = []
     for facet in topology[-2]:
         if len(facet) == 3:
-            edge_vertices += [[facet[i] for i in edge] for edge in triangle_edges]
+            edge_vertices += [[[facet[i] for i in edge] for edge in triangle_edges]]
         elif len(facet) == 4:
-            edge_vertices += [[facet[i] for i in edge] for edge in quadrilateral_edges]
+            edge_vertices += [[[facet[i] for i in edge] for edge in quadrilateral_edges]]
         else:
             raise ValueError("Only triangular and quadrilateral faces supported.")
 
-    out = numpy.array(edge_vertices, dtype=int)
-    return L.ArrayDecl("static const unsigned int", f"{cellname}_{tablename}", out.shape, out)
+    out = np.array(edge_vertices, dtype=int)
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.INT)
+    return L.ArrayDecl(symbol, values=out, const=True)
 
 
-def reference_facet_jacobian(L, tablename, cellname, type: str):
+def reference_facet_jacobian(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     out = basix.cell.facet_jacobians(celltype)
-    return L.ArrayDecl(f"static const {type}", f"{cellname}_{tablename}", out.shape, out)
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.REAL)
+    return L.ArrayDecl(symbol, values=out, const=True)
 
 
-def reference_cell_volume(L, tablename, cellname, type: str):
+def reference_cell_volume(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     out = basix.cell.volume(celltype)
-    return L.VariableDecl(f"static const {type}", f"{cellname}_{tablename}", out)
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.REAL)
+    return L.VariableDecl(symbol, out)
 
 
-def reference_facet_volume(L, tablename, cellname, type: str):
+def reference_facet_volume(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     volumes = basix.cell.facet_reference_volumes(celltype)
     for i in volumes[1:]:
-        if not numpy.isclose(i, volumes[0]):
+        if not np.isclose(i, volumes[0]):
             raise ValueError("Reference facet volume not supported for this cell type.")
-    return L.VariableDecl(f"static const {type}", f"{cellname}_{tablename}", volumes[0])
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.REAL)
+    return L.VariableDecl(symbol, volumes[0])
 
 
-def reference_edge_vectors(L, tablename, cellname, type: str):
+def reference_edge_vectors(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     topology = basix.topology(celltype)
     geometry = basix.geometry(celltype)
     edge_vectors = [geometry[j] - geometry[i] for i, j in topology[1]]
-    out = numpy.array(edge_vectors[cellname])
-    return L.ArrayDecl(f"static const {type}", f"{cellname}_{tablename}", out.shape, out)
+    out = np.array(edge_vectors)
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.REAL)
+    return L.ArrayDecl(symbol, values=out, const=True)
 
 
-def facet_reference_edge_vectors(L, tablename, cellname, type: str):
+def facet_reference_edge_vectors(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     topology = basix.topology(celltype)
     geometry = basix.geometry(celltype)
@@ -100,17 +105,20 @@ def facet_reference_edge_vectors(L, tablename, cellname, type: str):
         else:
             raise ValueError("Only triangular and quadrilateral faces supported.")
 
-    out = numpy.array(edge_vectors)
-    return L.ArrayDecl(f"static const {type}", f"{cellname}_{tablename}", out.shape, out)
+    out = np.array(edge_vectors)
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.REAL)
+    return L.ArrayDecl(symbol, values=out, const=True)
 
 
-def reference_facet_normals(L, tablename, cellname, type: str):
+def reference_facet_normals(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     out = basix.cell.facet_outward_normals(celltype)
-    return L.ArrayDecl(f"static const {type}", f"{cellname}_{tablename}", out.shape, out)
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.REAL)
+    return L.ArrayDecl(symbol, values=out, const=True)
 
 
-def facet_orientation(L, tablename, cellname, type: str):
+def facet_orientation(tablename, cellname):
     celltype = getattr(basix.CellType, cellname)
     out = basix.cell.facet_orientations(celltype)
-    return L.ArrayDecl(f"static const {type}", f"{cellname}_{tablename}", out.shape, out)
+    symbol = L.Symbol(f"{cellname}_{tablename}", dtype=L.DataType.REAL)
+    return L.ArrayDecl(symbol, values=out, const=True)
