@@ -1,4 +1,4 @@
-# Copyright (C) 2015-2021 Martin Sandve Alnæs, Michal Habera, Igor Baratta
+# Copyright (C) 2015-2023 Martin Sandve Alnæs, Michal Habera, Igor Baratta, Chris Richardson
 #
 # This file is part of FFCx. (https://www.fenicsproject.org)
 #
@@ -8,93 +8,15 @@ import collections
 import logging
 from typing import Any, Dict, List, Set, Tuple
 
-import numpy
-
 import ufl
 from ffcx.codegeneration import geometry
-from ffcx.codegeneration import integrals_template as ufcx_integrals
-from ffcx.codegeneration.backend import FFCXBackend
-from ffcx.codegeneration.C.cnodes import BinOp, CNode
-from ffcx.codegeneration.C.format_lines import format_indented_lines
+from ffcx.ir.elementtables import piecewise_ttypes
 from ffcx.ir.integral import BlockDataT
+import ffcx.codegeneration.lnodes as L
+from ffcx.codegeneration.lnodes import LNode, BinOp
 from ffcx.ir.representationutils import QuadratureRule
-from ffcx.naming import cdtype_to_numpy, scalar_to_value_type
 
 logger = logging.getLogger("ffcx")
-
-
-def generator(ir, options):
-    logger.info("Generating code for integral:")
-    logger.info(f"--- type: {ir.integral_type}")
-    logger.info(f"--- name: {ir.name}")
-
-    """Generate code for an integral."""
-    factory_name = ir.name
-
-    # Format declaration
-    declaration = ufcx_integrals.declaration.format(factory_name=factory_name)
-
-    # Create FFCx C backend
-    backend = FFCXBackend(ir, options)
-
-    # Configure kernel generator
-    ig = IntegralGenerator(ir, backend)
-
-    # Generate code ast for the tabulate_tensor body
-    parts = ig.generate()
-
-    # Format code as string
-    body = format_indented_lines(parts.cs_format(ir.precision), 1)
-
-    # Generate generic FFCx code snippets and add specific parts
-    code = {}
-    code["class_type"] = ir.integral_type + "_integral"
-    code["name"] = ir.name
-    code["members"] = ""
-    code["constructor"] = ""
-    code["constructor_arguments"] = ""
-    code["initializer_list"] = ""
-    code["destructor"] = ""
-
-    L = backend.language
-    if len(ir.enabled_coefficients) > 0:
-        code["enabled_coefficients_init"] = L.ArrayDecl(
-            "bool", f"enabled_coefficients_{ir.name}", values=ir.enabled_coefficients,
-            sizes=len(ir.enabled_coefficients))
-        code["enabled_coefficients"] = f"enabled_coefficients_{ir.name}"
-    else:
-        code["enabled_coefficients_init"] = ""
-        code["enabled_coefficients"] = L.Null()
-
-    code["additional_includes_set"] = set()  # FIXME: Get this out of code[]
-
-    #tabulate_tensor_code = ""
-
-    if ir.integral_type in ufl.custom_integral_types:
-        code["tabulate_tensor"] = ""
-        code["custom_tabulate_tensor"] = body
-        #tabulate_tensor_code = code["custom_tabulate_tensor"]
-    else:
-        code["tabulate_tensor"] = body
-        code["custom_tabulate_tensor"] = ""
-        #tabulate_tensor_code = code["tabulate_tensor"]
-
-    if options["tabulate_tensor_void"]:
-        code["tabulate_tensor"] = ""
-
-    implementation = ufcx_integrals.factory.format(
-        factory_name=factory_name,
-        enabled_coefficients=code["enabled_coefficients"],
-        enabled_coefficients_init=code["enabled_coefficients_init"],
-        tabulate_tensor=code["tabulate_tensor"],
-        custom_tabulate_tensor=code["custom_tabulate_tensor"],
-        needs_facet_permutations="true" if ir.needs_facet_permutations else "false",
-        scalar_type=options["scalar_type"],
-        geom_type=scalar_to_value_type(options["scalar_type"]),
-        np_scalar_type=cdtype_to_numpy(options["scalar_type"]),
-        coordinate_element=L.AddressOf(L.Symbol(ir.coordinate_element)))
-
-    return declaration, implementation
 
 
 class IntegralGenerator(object):
@@ -103,7 +25,6 @@ class IntegralGenerator(object):
         self.ir = ir
 
         # Backend specific plugin with attributes
-        # - language: for translating ufl operators to target language
         # - symbols: for translating ufl operators to target language
         # - definitions: for defining backend specific variables
         # - access: for accessing backend specific variables
@@ -135,7 +56,7 @@ class IntegralGenerator(object):
         Scope is determined by quadrature_rule which identifies the
         quadrature loop scope or None if outside quadrature loops.
 
-        v is the ufl expression and vaccess is the CNodes
+        v is the ufl expression and vaccess is the LNodes
         expression to access the value in the code.
 
         """
@@ -150,10 +71,10 @@ class IntegralGenerator(object):
         If v is not found in quadrature loop scope, the piecewise
         scope (None) is checked.
 
-        Returns the CNodes expression to access the value in the code.
+        Returns the LNodes expression to access the value in the code.
         """
         if v._ufl_is_literal_:
-            return self.backend.ufl_to_language.get(v)
+            return L.ufl_to_lnodes(v)
         f = self.scopes[quadrature_rule].get(v)
         if f is None:
             f = self.scopes[None].get(v)
@@ -161,13 +82,12 @@ class IntegralGenerator(object):
 
     def new_temp_symbol(self, basename):
         """Create a new code symbol named basename + running counter."""
-        L = self.backend.language
         name = "%s%d" % (basename, self.symbol_counters[basename])
         self.symbol_counters[basename] += 1
-        return L.Symbol(name)
+        return L.Symbol(name, dtype=L.DataType.SCALAR)
 
     def get_temp_symbol(self, tempname, key):
-        key = (tempname, ) + key
+        key = (tempname,) + key
         s = self.shared_symbols.get(key)
         defined = s is not None
         if not defined:
@@ -182,32 +102,21 @@ class IntegralGenerator(object):
         context that matches a suitable version of the UFC
         tabulate_tensor signatures.
         """
-        L = self.backend.language
-
         # Assert that scopes are empty: expecting this to be called only
         # once
         assert not any(d for d in self.scopes.values())
 
         parts = []
-        scalar_type = self.backend.access.options["scalar_type"]
-        value_type = scalar_to_value_type(scalar_type)
-        alignment = self.ir.options['assume_aligned']
-        if alignment != -1:
-            scalar_type = self.backend.access.options["scalar_type"]
-            parts += [L.VerbatimStatement(f"A = ({scalar_type}*)__builtin_assume_aligned(A, {alignment});"),
-                      L.VerbatimStatement(f"w = (const {scalar_type}*)__builtin_assume_aligned(w, {alignment});"),
-                      L.VerbatimStatement(f"c = (const {scalar_type}*)__builtin_assume_aligned(c, {alignment});"),
-                      L.VerbatimStatement(f"coordinate_dofs = (const {value_type}*)__builtin_assume_aligned(coordinate_dofs, {alignment});")]  # noqa
 
         # Generate the tables of quadrature points and weights
-        parts += self.generate_quadrature_tables(value_type)
+        parts += self.generate_quadrature_tables()
 
         # Generate the tables of basis function values and
         # pre-integrated blocks
-        parts += self.generate_element_tables(value_type)
+        parts += self.generate_element_tables()
 
         # Generate the tables of geometry data that are needed
-        parts += self.generate_geometry_tables(value_type)
+        parts += self.generate_geometry_tables()
 
         # Loop generation code will produce parts to go before
         # quadloops, to define the quadloops, and to go after the
@@ -225,7 +134,6 @@ class IntegralGenerator(object):
             # Generate code to integrate reusable blocks of final
             # element tensor
             pre_definitions, preparts, quadparts = self.generate_quadrature_loop(rule)
-
             all_preparts += preparts
             all_quadparts += quadparts
             all_predefinitions.update(pre_definitions)
@@ -239,11 +147,9 @@ class IntegralGenerator(object):
 
         return L.StatementList(parts)
 
-    def generate_quadrature_tables(self, value_type: str) -> List[str]:
+    def generate_quadrature_tables(self):
         """Generate static tables of quadrature points and weights."""
-        L = self.backend.language
-
-        parts: List[str] = []
+        parts = []
 
         # No quadrature tables for custom (given argument) or point
         # (evaluation in single vertex)
@@ -251,26 +157,18 @@ class IntegralGenerator(object):
         if self.ir.integral_type in skip:
             return parts
 
-        padlen = self.ir.options["padlen"]
-
         # Loop over quadrature rules
         for quadrature_rule, integrand in self.ir.integrand.items():
-
-            num_points = quadrature_rule.weights.shape[0]
-
             # Generate quadrature weights array
             wsym = self.backend.symbols.weights_table(quadrature_rule)
-            parts += [L.ArrayDecl(f"static const {value_type}", wsym, num_points,
-                                  quadrature_rule.weights, padlen=padlen)]
+            parts += [L.ArrayDecl(wsym, values=quadrature_rule.weights, const=True)]
 
         # Add leading comment if there are any tables
         parts = L.commented_code_list(parts, "Quadrature rules")
         return parts
 
-    def generate_geometry_tables(self, float_type: str):
+    def generate_geometry_tables(self):
         """Generate static tables of geometry data."""
-        L = self.backend.language
-
         ufl_geometry = {
             ufl.geometry.FacetEdgeVectors: "facet_edge_vertices",
             ufl.geometry.CellFacetJacobian: "reference_facet_jacobian",
@@ -294,175 +192,45 @@ class IntegralGenerator(object):
         parts = []
         for i, cell_list in cells.items():
             for c in cell_list:
-                parts.append(geometry.write_table(L, ufl_geometry[i], c, float_type))
+                parts.append(geometry.write_table(ufl_geometry[i], c))
 
         return parts
 
-    def generate_element_tables(self, float_type: str):
+    def generate_element_tables(self):
         """Generate static tables with precomputed element basisfunction values in quadrature points."""
-        L = self.backend.language
         parts = []
         tables = self.ir.unique_tables
-        padlen = self.ir.options["padlen"]
-
-        table_names = sorted(tables)
-
+        table_types = self.ir.unique_table_types
         if self.ir.integral_type in ufl.custom_integral_types:
-            element_ids = self.ir.element_ids
-            element_names = self.ir.element_names
-            element_tables = self.ir.unique_element_tables
-            element_deriv_order = self.ir.element_deriv_order
-            el_def_part, tab_part, copy_tab_part = self.generate_custom_integral_tables(table_names, tables,
-                                                                                     float_type, element_ids,
-                                                                                     element_names, element_deriv_order,
-                                                                                     element_tables)
-            parts += el_def_part
-            parts += tab_part
-            parts += copy_tab_part
+            # Define only piecewise tables
+            table_names = [name for name in sorted(tables) if table_types[name] in piecewise_ttypes]
         else:
             # Define all tables
-            for name in table_names:
-                table = tables[name]
-                parts += self.declare_table(name, table, padlen, float_type)
+            table_names = sorted(tables)
 
-            # Add leading comment if there are any tables
-            parts = L.commented_code_list(parts, [
-                "Precomputed values of basis functions and precomputations",
-                "FE* dimensions: [permutation][entities][points][dofs]"])
-        return parts
-
-    def generate_custom_integral_tables(self, table_names, tables, float_type, element_ids,
-                                        element_names, element_deriv_order, element_tables):
-        L = self.backend.language
-        element_def_parts = []
-
-        for id, e in element_tables.items():
-            component_element, offset, stride = e.element.get_component_element(e.fc)
-            decl = "// Represented element is " + element_names[e.element] + " component " + str(e.fc) + "\n"
-            decl += "basix_element* basix_element_" + str(id) + " = basix_element_create("
-            decl += str(int(component_element.element_family)) + ", "
-            decl += str(int(component_element.cell_type)) + ", "
-            decl += str(component_element.degree()) + ", "
-            decl += str(int(component_element.lagrange_variant)) + ", "
-            decl += str(int(component_element.dpc_variant)) + ", "
-            decl+= "true" if component_element.discontinuous else "false"
-            decl += ");"
-
-            element_def_parts += [L.VerbatimStatement(decl)]
-        comment = "FIXME: basix elements should be created in seperate function for efficiency"
-        element_def_parts = L.commented_code_list(element_def_parts,
-                                        [comment])
-
-        tabulate_parts = []
-
-        for id, e in element_tables.items():
-            component_element, offset, stride = e.element.get_component_element(e.fc)
-            nd = e.deriv_order
-
-            cshape_str = "shape_" + str(id)
-            decl = "int " + cshape_str + "[4];\n"
-            decl += "basix_element_tabulate_shape("
-            decl += "basix_element_" + str(id) + ", "
-            decl += str(self.backend.symbols.custom_num_points()) + ", "
-            decl += str(nd) + ", "
-            decl += cshape_str + ");\n"
-
-            # int table_size = shape[0]*shape[1]*shape[2]*shape[3];
-            # double table[shape[0]][shape[1]][shape[2]][shape[3]];
-            # basix_element_tabulate(element, points, num_points, nd, (double*) table, table_size);
-            decl += "int table_size_" + str(id) + "="
-            decl += cshape_str + "[0]*" + cshape_str + "[1]*" + cshape_str + "[2]*" + cshape_str + "[3];\n"
-            decl += "double table_" + str(id) + "[" + cshape_str + "[0]][" + cshape_str + "[1]]["
-            decl += cshape_str + "[2]][" + cshape_str + "[3]];\n"
-
-            decl += "basix_element_tabulate("
-            decl += "basix_element_" + str(id) + ", "
-            decl += str(e.gdim) + ", "
-            decl += "points, "
-            decl += str(self.backend.symbols.custom_num_points()) + ", "
-            decl += str(nd) + ", "
-            decl += "(double *) table_"+ str(id) + ", table_size_" + str(id) + ");\n"
-            decl += "basix_element_destroy(basix_element_" + str(id) + ");"
-
-            tabulate_parts += [L.VerbatimStatement(decl)]
-
-        copy_table_parts = []
-
-        for id, e in element_tables.items():
-            name = e.name
-            #table for standard integrals
+        for name in table_names:
             table = tables[name]
-
-            # Replace number of points in array with symbolic number of points that is passed to
-            # tabulate tensor at run-time
-            brackets = ''.join("[%d]" % table.shape[0])
-            brackets += ''.join("[%d]" % table.shape[1])
-            brackets += ''.join("[%s]" % self.backend.symbols.custom_num_points())
-            brackets += ''.join("[%d]" % table.shape[3])
-
-            # Declare array of the type e.g. double FE#_C#[1][1][num_points][3];
-            # to be filled at run-time using basix tabulate call
-            decl = float_type + " " + name + brackets + "; \n"
-
-            copy_table_parts += [L.VerbatimStatement(decl)]
-
-            num_points = self.backend.symbols.custom_num_points()
-            iq = self.backend.symbols.quadrature_loop_index()
-            arg_indices = tuple(self.backend.symbols.argument_loop_index(i) for i in range(3))
-
-            brackets = []
-            if (table.shape[0] > 1):
-                brackets = ''.join("[%s]" % arg_indices[0])
-            else:
-                brackets = ''.join("[0]")
-
-            if (table.shape[1] > 1):
-                brackets += ''.join("[%s]" % arg_indices[1])
-            else:
-                brackets += ''.join("[0]")
-
-            brackets += ''.join("[%s]" % iq)
-            brackets += ''.join("[%s]" % arg_indices[2])
-            body = name + brackets + " = "
-
-            # Element tables gives metadata associated to name: (id, basix_index,fc)
-            # table dimensions: [derivatives (basix::indexing)][point index][basis function index][function component]
-            brackets = ''.join("[%d]" %  e.basix_index)
-            brackets += ''.join("[%s]" % iq)
-            brackets += ''.join("[%s]" % arg_indices[2])
-            brackets += ''.join("[%d]" % e.fc)
-            body += "table_" + str(id) + brackets + ";"
-
-            body = [L.ForRange(arg_indices[2], 0, table.shape[3], body=body)]
-            body = [L.ForRange(iq, 0, num_points, body=body)]
-            if (table.shape[1] > 1):
-                body = [L.ForRange(arg_indices[1], 0, table.shape[1], body=body)]
-            if (table.shape[0] > 1):
-                body = [L.ForRange(arg_indices[0], 0, table.shape[0], body=body)]
-
-            copy_table_parts += body
+            parts += self.declare_table(name, table)
 
         # Add leading comment if there are any tables
-        copy_table_parts = L.commented_code_list(copy_table_parts, [
-            "Array for basis evaluations and basis derivative evaluations",
+        parts = L.commented_code_list(parts, [
+            "Precomputed values of basis functions and precomputations",
             "FE* dimensions: [permutation][entities][points][dofs]"])
+        return parts
 
-        return element_def_parts, tabulate_parts, copy_table_parts
-
-    def declare_table(self, name, table, padlen, value_type: str):
+    def declare_table(self, name, table):
         """Declare a table.
 
         If the dof dimensions of the table have dof rotations, apply
         these rotations.
 
         """
-        L = self.backend.language
-
-        return [L.ArrayDecl(f"static const {value_type}", name, table.shape, table, padlen=padlen)]
+        table_symbol = L.Symbol(name, dtype=L.DataType.REAL)
+        self.backend.symbols.element_tables[name] = table_symbol
+        return [L.ArrayDecl(table_symbol, values=table, const=True)]
 
     def generate_quadrature_loop(self, quadrature_rule: QuadratureRule):
         """Generate quadrature loop with for this quadrature_rule."""
-        L = self.backend.language
         # Generate varying partition
         pre_definitions, body = self.generate_varying_partition(quadrature_rule)
 
@@ -480,21 +248,16 @@ class IntegralGenerator(object):
             quadparts = []
         else:
             num_points = quadrature_rule.points.shape[0]
-            # Override num_points with string for custom integral
-            if self.ir.integral_type in ufl.custom_integral_types:
-                num_points = self.backend.symbols.custom_num_points()
-            iq = self.backend.symbols.quadrature_loop_index()
+            iq = self.backend.symbols.quadrature_loop_index
             quadparts = [L.ForRange(iq, 0, num_points, body=body)]
 
         return pre_definitions, preparts, quadparts
 
     def generate_piecewise_partition(self, quadrature_rule):
-        L = self.backend.language
-
         # Get annotated graph of factorisation
         F = self.ir.integrand[quadrature_rule]["factorization"]
 
-        arraysymbol = L.Symbol(f"sp_{quadrature_rule.id()}")
+        arraysymbol = L.Symbol(f"sp_{quadrature_rule.id()}", dtype=L.DataType.SCALAR)
         pre_definitions, parts = self.generate_partition(arraysymbol, F, "piecewise", None)
         assert len(pre_definitions) == 0, "Quadrature independent code should have not pre-definitions"
         parts = L.commented_code_list(
@@ -503,19 +266,17 @@ class IntegralGenerator(object):
         return parts
 
     def generate_varying_partition(self, quadrature_rule):
-        L = self.backend.language
 
         # Get annotated graph of factorisation
         F = self.ir.integrand[quadrature_rule]["factorization"]
 
-        arraysymbol = L.Symbol(f"sv_{quadrature_rule.id()}")
+        arraysymbol = L.Symbol(f"sv_{quadrature_rule.id()}", dtype=L.DataType.SCALAR)
         pre_definitions, parts = self.generate_partition(arraysymbol, F, "varying", quadrature_rule)
         parts = L.commented_code_list(parts, f"Varying computations for quadrature rule {quadrature_rule.id()}")
 
         return pre_definitions, parts
 
     def generate_partition(self, symbol, F, mode, quadrature_rule):
-        L = self.backend.language
 
         definitions = dict()
         pre_definitions = dict()
@@ -524,7 +285,6 @@ class IntegralGenerator(object):
         use_symbol_array = True
 
         for i, attr in F.nodes.items():
-
             if attr['status'] != mode:
                 continue
             v = attr['expression']
@@ -534,12 +294,13 @@ class IntegralGenerator(object):
             # cache
             if not self.get_var(quadrature_rule, v):
                 if v._ufl_is_literal_:
-                    vaccess = self.backend.ufl_to_language.get(v)
+                    vaccess = L.ufl_to_lnodes(v)
                 elif mt is not None:
                     # All finite element based terminals have table
                     # data, as well as some, but not all, of the
                     # symbolic geometric terminals
                     tabledata = attr.get('tr')
+
                     # Backend specific modified terminal translation
                     vaccess = self.backend.access.get(mt.terminal, mt, tabledata, quadrature_rule)
                     predef, vdef = self.backend.definitions.get(mt.terminal, mt, tabledata, quadrature_rule, vaccess)
@@ -563,7 +324,7 @@ class IntegralGenerator(object):
 
                     # Mapping UFL operator to target language
                     self._ufl_names.add(v._ufl_handler_name_)
-                    vexpr = self.backend.ufl_to_language.get(v, *vops)
+                    vexpr = L.ufl_to_lnodes(v, *vops)
 
                     # Create a new intermediate for each subexpression
                     # except boolean conditions and its childs
@@ -590,9 +351,8 @@ class IntegralGenerator(object):
                             vaccess = symbol[j]
                             intermediates.append(L.Assign(vaccess, vexpr))
                         else:
-                            scalar_type = self.backend.access.options["scalar_type"]
                             vaccess = L.Symbol("%s_%d" % (symbol.name, j))
-                            intermediates.append(L.VariableDecl(f"const {scalar_type}", vaccess, vexpr))
+                            intermediates.append(L.VariableDecl(vaccess, vexpr))
 
                 # Store access node for future reference
                 self.set_var(quadrature_rule, v, vaccess)
@@ -604,9 +364,7 @@ class IntegralGenerator(object):
 
         if intermediates:
             if use_symbol_array:
-                padlen = self.ir.options["padlen"]
-                parts += [L.ArrayDecl(self.backend.access.options["scalar_type"],
-                                      symbol, len(intermediates), padlen=padlen)]
+                parts += [L.ArrayDecl(symbol, sizes=len(intermediates))]
             parts += intermediates
         return pre_definitions, parts
 
@@ -678,11 +436,9 @@ class IntegralGenerator(object):
         Should be called with quadrature_rule=None for
         quadloop-independent blocks.
         """
-        L = self.backend.language
-
         # The parts to return
-        preparts: List[CNode] = []
-        quadparts: List[CNode] = []
+        preparts: List[LNode] = []
+        quadparts: List[LNode] = []
 
         # RHS expressions grouped by LHS "dofmap"
         rhs_expressions = collections.defaultdict(list)
@@ -690,7 +446,7 @@ class IntegralGenerator(object):
         block_rank = len(blockmap)
         blockdims = tuple(len(dofmap) for dofmap in blockmap)
 
-        iq = self.backend.symbols.quadrature_loop_index()
+        iq = self.backend.symbols.quadrature_loop_index
 
         # Override dof index with quadrature loop index for arguments
         # with quadrature element, to index B like B[iq*num_dofs + iq]
@@ -719,7 +475,7 @@ class IntegralGenerator(object):
 
             # Quadrature weight was removed in representation, add it back now
             if self.ir.integral_type in ufl.custom_integral_types:
-                weights = self.backend.symbols.custom_weights_table()
+                weights = self.backend.symbols.custom_weights_table
                 weight = weights[iq]
             else:
                 weights = self.backend.symbols.weights_table(quadrature_rule)
@@ -734,14 +490,9 @@ class IntegralGenerator(object):
                 key = (quadrature_rule, factor_index, blockdata.all_factors_piecewise)
                 fw, defined = self.get_temp_symbol("fw", key)
                 if not defined:
-                    scalar_type = self.backend.access.options["scalar_type"]
-                    quadparts.append(L.VariableDecl(f"const {scalar_type}", fw, fw_rhs))
+                    quadparts.append(L.VariableDecl(fw, fw_rhs))
 
             assert not blockdata.transposed, "Not handled yet"
-            A_shape = self.ir.tensor_shape
-
-            Asym = self.backend.symbols.element_tensor()
-            A = L.FlattenedArray(Asym, dims=A_shape)
 
             # Fetch code to access modified arguments
             arg_factors = self.get_arg_factors(blockdata, block_rank, quadrature_rule, iq, B_indices)
@@ -762,7 +513,7 @@ class IntegralGenerator(object):
         # List of statements to keep in the inner loop
         keep = collections.defaultdict(list)
         # List of temporary array declarations
-        pre_loop: List[CNode] = []
+        pre_loop: List[LNode] = []
         # List of loop invariant expressions to hoist
         hoist: List[BinOp] = []
 
@@ -788,34 +539,32 @@ class IntegralGenerator(object):
                 # floating point operations (factorize expressions by
                 # grouping)
                 for statement in hoist_rhs:
-                    sum = []
-                    for rhs in hoist_rhs[statement]:
-                        sum.append(L.float_product(rhs))
-                    sum = L.Sum(sum)
+                    sum = L.Sum([L.float_product(rhs) for rhs in hoist_rhs[statement]])
 
                     lhs = None
                     for h in hoist:
-                        if (h.rhs == sum):
+                        if h.rhs == sum:
                             lhs = h.lhs
                             break
                     if lhs:
                         keep[indices].append(L.float_product([statement, lhs]))
                     else:
                         t = self.new_temp_symbol("t")
-                        scalar_type = self.backend.access.options["scalar_type"]
-                        pre_loop.append(L.ArrayDecl(scalar_type, t, blockdims[0]))
+                        pre_loop.append(L.ArrayDecl(t, sizes=blockdims[0]))
                         keep[indices].append(L.float_product([statement, t[B_indices[0]]]))
                         hoist.append(L.Assign(t[B_indices[i - 1]], sum))
             else:
                 keep[indices] = rhs_expressions[indices]
 
-        hoist_code: List[CNode] = [L.ForRange(B_indices[0], 0, blockdims[0], body=hoist)] if hoist else []
+        hoist_code: List[LNode] = [L.ForRange(B_indices[0], 0, blockdims[0], body=hoist)] if hoist else []
 
-        body: List[CNode] = []
+        body: List[LNode] = []
 
+        A = self.backend.symbols.element_tensor
+        A_shape = self.ir.tensor_shape
         for indices in keep:
-            sum = L.Sum(keep[indices])
-            body.append(L.AssignAdd(A[indices], sum))
+            multi_index = L.MultiIndex(list(indices), A_shape)
+            body.append(L.AssignAdd(A[multi_index], L.Sum(keep[indices])))
 
         for i in reversed(range(block_rank)):
             body = [L.ForRange(B_indices[i], 0, blockdims[i], body=body)]
@@ -837,8 +586,6 @@ class IntegralGenerator(object):
         determine how many loops should fuse at a time.
 
         """
-        L = self.backend.language
-
         loops = collections.defaultdict(list)
         pre_loop = []
         for access, definition in definitions.items():
