@@ -8,10 +8,10 @@
 import logging
 import typing
 
-import numpy
+import numpy as np
+import numpy.typing as npt
 
 import ufl
-import ufl.utils.derivativetuples
 from ffcx.element_interface import (QuadratureElement, basix_index,
                                     convert_element)
 from ffcx.ir.representationutils import (create_quadrature_points_and_weights,
@@ -20,7 +20,7 @@ from ffcx.ir.representationutils import (create_quadrature_points_and_weights,
 
 logger = logging.getLogger("ffcx")
 
-# Using same defaults as numpy.allclose
+# Using same defaults as np.allclose
 default_rtol = 1e-6
 default_atol = 1e-9
 
@@ -31,13 +31,13 @@ uniform_ttypes = ("fixed", "ones", "zeros", "uniform")
 class ModifiedTerminalElement(typing.NamedTuple):
     element: ufl.FiniteElementBase
     averaged: str
-    local_derivatives: typing.Tuple[int]
+    local_derivatives: typing.Tuple[int, ...]
     fc: int
 
 
 class UniqueTableReferenceT(typing.NamedTuple):
     name: str
-    values: numpy.typing.NDArray[numpy.float64]
+    values: npt.NDArray[np.float64]
     offset: int
     block_size: int
     ttype: str
@@ -60,12 +60,12 @@ class ElementTables(typing.NamedTuple):
 
 
 def equal_tables(a, b, rtol=default_rtol, atol=default_atol):
-    a = numpy.asarray(a)
-    b = numpy.asarray(b)
+    a = np.asarray(a)
+    b = np.asarray(b)
     if a.shape != b.shape:
         return False
     else:
-        return numpy.allclose(a, b, rtol=rtol, atol=atol)
+        return np.allclose(a, b, rtol=rtol, atol=atol)
 
 
 def clamp_table_small_numbers(table,
@@ -74,9 +74,9 @@ def clamp_table_small_numbers(table,
                               numbers=(-1.0, 0.0, 1.0)):
     """Clamp almost 0,1,-1 values to integers. Returns new table."""
     # Get shape of table and number of columns, defined as the last axis
-    table = numpy.asarray(table)
+    table = np.asarray(table)
     for n in numbers:
-        table[numpy.where(numpy.isclose(table, n, rtol=rtol, atol=atol))] = n
+        table[np.where(np.isclose(table, n, rtol=rtol, atol=atol))] = n
     return table
 
 
@@ -122,12 +122,12 @@ def get_ffcx_table_values(points, cell, integral_type, element, avg, entitytype,
         else:
             # Make quadrature rule and get points and weights
             points, weights = create_quadrature_points_and_weights(
-                integral_type, cell, element.highest_degree(), "default")
+                integral_type, cell, element.highest_degree(), "default", [element])
 
     # Tabulate table of basis functions and derivatives in points for each entity
     tdim = cell.topological_dimension()
     entity_dim = integral_type_to_entity_dim(integral_type, tdim)
-    num_entities = ufl.cell.num_cell_entities[cell.cellname()][entity_dim]
+    num_entities = cell.num_sub_entities(entity_dim)
 
     # Extract arrays for the right scalar component
     component_tables = []
@@ -144,8 +144,8 @@ def get_ffcx_table_values(points, cell, integral_type, element, avg, entitytype,
         wsum = sum(weights)
         for entity, tbl in enumerate(component_tables):
             num_dofs = tbl.shape[1]
-            tbl = numpy.dot(tbl, weights) / wsum
-            tbl = numpy.reshape(tbl, (1, num_dofs))
+            tbl = np.dot(tbl, weights) / wsum
+            tbl = np.reshape(tbl, (1, num_dofs))
             component_tables[entity] = tbl
 
     # Loop over entities and fill table blockwise (each block = points x dofs)
@@ -153,8 +153,7 @@ def get_ffcx_table_values(points, cell, integral_type, element, avg, entitytype,
     assert len(component_tables) == num_entities
     num_points, num_dofs = component_tables[0].shape
     shape = (1, num_entities, num_points, num_dofs)
-
-    res = numpy.zeros(shape)
+    res = np.zeros(shape)
     for entity in range(num_entities):
         res[:, entity, :, :] = component_tables[entity]
 
@@ -241,9 +240,8 @@ def get_modified_terminal_element(mt) -> typing.Optional[ModifiedTerminalElement
 
     assert (mt.averaged is None) or not (ld or gd)
     # Change derivatives format for table lookup
-    gdim = domain.geometric_dimension()
-    local_derivatives = ufl.utils.derivativetuples.derivative_listing_to_counts(
-        ld, gdim)
+    tdim = domain.topological_dimension()
+    local_derivatives: typing.Tuple[int, ...] = tuple(ld.count(i) for i in range(tdim))
 
     return ModifiedTerminalElement(element, mt.averaged, local_derivatives, fc)
 
@@ -251,8 +249,8 @@ def get_modified_terminal_element(mt) -> typing.Optional[ModifiedTerminalElement
 def permute_quadrature_interval(points, reflections=0):
     output = points.copy()
     for p in output:
-        assert len(p) < 2 or numpy.isclose(p[1], 0)
-        assert len(p) < 3 or numpy.isclose(p[2], 0)
+        assert len(p) < 2 or np.isclose(p[1], 0)
+        assert len(p) < 3 or np.isclose(p[2], 0)
     for i in range(reflections):
         for n, p in enumerate(output):
             output[n] = [1 - p[0]]
@@ -262,7 +260,7 @@ def permute_quadrature_interval(points, reflections=0):
 def permute_quadrature_triangle(points, reflections=0, rotations=0):
     output = points.copy()
     for p in output:
-        assert len(p) < 3 or numpy.isclose(p[2], 0)
+        assert len(p) < 3 or np.isclose(p[2], 0)
     for i in range(rotations):
         for n, p in enumerate(output):
             output[n] = [p[1], 1 - p[0] - p[1]]
@@ -275,7 +273,7 @@ def permute_quadrature_triangle(points, reflections=0, rotations=0):
 def permute_quadrature_quadrilateral(points, reflections=0, rotations=0):
     output = points.copy()
     for p in output:
-        assert len(p) < 3 or numpy.isclose(p[2], 0)
+        assert len(p) < 3 or np.isclose(p[2], 0)
     for i in range(rotations):
         for n, p in enumerate(output):
             output[n] = [p[1], 1 - p[0]]
@@ -347,7 +345,7 @@ def build_optimized_tables(quadrature_rule, cell, integral_type, entitytype,
                         integral_type, element, avg, entitytype, local_derivatives, flat_component))
 
                 t = new_table[0]
-                t['array'] = numpy.vstack([td['array'] for td in new_table])
+                t['array'] = np.vstack([td['array'] for td in new_table])
             elif tdim == 3:
                 cell_type = cell.cellname()
                 if cell_type == "tetrahedron":
@@ -360,7 +358,7 @@ def build_optimized_tables(quadrature_rule, cell, integral_type, entitytype,
                                 cell, integral_type, element, avg, entitytype, local_derivatives,
                                 flat_component))
                     t = new_table[0]
-                    t['array'] = numpy.vstack([td['array'] for td in new_table])
+                    t['array'] = np.vstack([td['array'] for td in new_table])
                 elif cell_type == "hexahedron":
                     new_table = []
                     for rot in range(4):
@@ -370,7 +368,7 @@ def build_optimized_tables(quadrature_rule, cell, integral_type, entitytype,
                                     quadrature_rule.points, ref, rot),
                                 cell, integral_type, element, avg, entitytype, local_derivatives, flat_component))
                     t = new_table[0]
-                    t['array'] = numpy.vstack([td['array'] for td in new_table])
+                    t['array'] = np.vstack([td['array'] for td in new_table])
         else:
             t = get_ffcx_table_values(quadrature_rule.points, cell,
                                       integral_type, element, avg, entitytype,
@@ -421,39 +419,39 @@ def build_optimized_tables(quadrature_rule, cell, integral_type, entitytype,
 
 
 def is_zeros_table(table, rtol=default_rtol, atol=default_atol):
-    return (numpy.product(table.shape) == 0
-            or numpy.allclose(table, numpy.zeros(table.shape), rtol=rtol, atol=atol))
+    return (np.prod(table.shape) == 0
+            or np.allclose(table, np.zeros(table.shape), rtol=rtol, atol=atol))
 
 
 def is_ones_table(table, rtol=default_rtol, atol=default_atol):
-    return numpy.allclose(table, numpy.ones(table.shape), rtol=rtol, atol=atol)
+    return np.allclose(table, np.ones(table.shape), rtol=rtol, atol=atol)
 
 
 def is_quadrature_table(table, rtol=default_rtol, atol=default_atol):
     _, num_entities, num_points, num_dofs = table.shape
-    Id = numpy.eye(num_points)
+    Id = np.eye(num_points)
     return (num_points == num_dofs and all(
-        numpy.allclose(table[0, i, :, :], Id, rtol=rtol, atol=atol) for i in range(num_entities)))
+        np.allclose(table[0, i, :, :], Id, rtol=rtol, atol=atol) for i in range(num_entities)))
 
 
 def is_permuted_table(table, rtol=default_rtol, atol=default_atol):
     return not all(
-        numpy.allclose(table[0, :, :, :],
-                       table[i, :, :, :], rtol=rtol, atol=atol)
+        np.allclose(table[0, :, :, :],
+                    table[i, :, :, :], rtol=rtol, atol=atol)
         for i in range(1, table.shape[0]))
 
 
 def is_piecewise_table(table, rtol=default_rtol, atol=default_atol):
     return all(
-        numpy.allclose(table[0, :, 0, :],
-                       table[0, :, i, :], rtol=rtol, atol=atol)
+        np.allclose(table[0, :, 0, :],
+                    table[0, :, i, :], rtol=rtol, atol=atol)
         for i in range(1, table.shape[2]))
 
 
 def is_uniform_table(table, rtol=default_rtol, atol=default_atol):
     return all(
-        numpy.allclose(table[0, 0, :, :],
-                       table[0, i, :, :], rtol=rtol, atol=atol)
+        np.allclose(table[0, 0, :, :],
+                    table[0, i, :, :], rtol=rtol, atol=atol)
         for i in range(1, table.shape[1]))
 
 
