@@ -8,8 +8,9 @@ import numpy as np
 import pytest
 
 import ffcx.codegeneration.jit
+import basix.ufl
 import ufl
-from ffcx.naming import cdtype_to_numpy, scalar_to_value_type
+from ffcx.codegeneration.utils import cdtype_to_numpy, scalar_to_value_type
 
 
 @pytest.mark.parametrize("mode",
@@ -21,9 +22,10 @@ from ffcx.naming import cdtype_to_numpy, scalar_to_value_type
                              "float _Complex"
                          ])
 def test_additive_facet_integral(mode, compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     a = ufl.inner(u, v) * ufl.ds
     forms = [a]
     compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(
@@ -35,11 +37,13 @@ def test_additive_facet_integral(mode, compile_args):
     ffi = module.ffi
     form0 = compiled_forms[0]
 
-    assert form0.num_integrals(module.lib.exterior_facet) == 1
-    ids = form0.integral_ids(module.lib.exterior_facet)
-    assert ids[0] == -1
+    integral_offsets = form0.form_integral_offsets
+    ex = module.lib.exterior_facet
+    assert integral_offsets[ex + 1] - integral_offsets[ex] == 1
+    integral_id = form0.form_integral_ids[integral_offsets[ex]]
+    assert integral_id == -1
 
-    default_integral = form0.integrals(module.lib.exterior_facet)[0]
+    default_integral = form0.form_integrals[integral_offsets[ex]]
 
     np_type = cdtype_to_numpy(mode)
     A = np.zeros((3, 3), dtype=np_type)
@@ -70,9 +74,10 @@ def test_additive_facet_integral(mode, compile_args):
 
 @pytest.mark.parametrize("mode", ["double", "float", "long double", "double _Complex", "float _Complex"])
 def test_additive_cell_integral(mode, compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     a = ufl.inner(ufl.grad(u), ufl.grad(v)) * ufl.dx
     forms = [a]
     compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(
@@ -84,11 +89,14 @@ def test_additive_cell_integral(mode, compile_args):
     ffi = module.ffi
     form0 = compiled_forms[0]
 
-    assert form0.num_integrals(module.lib.cell) == 1
-    ids = form0.integral_ids(module.lib.cell)
-    assert ids[0] == -1
+    cell = module.lib.cell
+    offsets = form0.form_integral_offsets
+    num_integrals = offsets[cell + 1] - offsets[cell]
+    assert num_integrals == 1
+    integral_id = form0.form_integral_ids[offsets[cell]]
+    assert integral_id == -1
 
-    default_integral = form0.integrals(0)[0]
+    default_integral = form0.form_integrals[offsets[cell]]
 
     np_type = cdtype_to_numpy(mode)
     A = np.zeros((3, 3), dtype=np_type)
