@@ -9,9 +9,10 @@ import pytest
 import sympy
 from sympy.abc import x, y, z
 
+import basix.ufl
 import ffcx.codegeneration.jit
 import ufl
-from ffcx.naming import cdtype_to_numpy, scalar_to_value_type
+from ffcx.codegeneration.utils import cdtype_to_numpy, scalar_to_value_type
 
 
 @pytest.mark.parametrize("mode,expected_result", [
@@ -23,10 +24,11 @@ from ffcx.naming import cdtype_to_numpy, scalar_to_value_type
          dtype=np.complex128)),
 ])
 def test_laplace_bilinear_form_2d(mode, expected_result, compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    kappa = ufl.Constant(cell, shape=(2, 2))
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    kappa = ufl.Constant(domain, shape=(2, 2))
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
 
     a = ufl.tr(kappa) * ufl.inner(ufl.grad(u), ufl.grad(v)) * ufl.dx
     forms = [a]
@@ -39,11 +41,13 @@ def test_laplace_bilinear_form_2d(mode, expected_result, compile_args):
     ffi = module.ffi
     form0 = compiled_forms[0]
 
-    assert form0.num_integrals(module.lib.cell) == 1
-    ids = form0.integral_ids(module.lib.cell)
-    assert ids[0] == -1
+    offsets = form0.form_integral_offsets
+    cell = module.lib.cell
+    assert offsets[cell + 1] - offsets[cell] == 1
+    integral_id = form0.form_integral_ids[offsets[cell]]
+    assert integral_id == -1
 
-    default_integral = form0.integrals(module.lib.cell)[0]
+    default_integral = form0.form_integrals[offsets[cell]]
 
     np_type = cdtype_to_numpy(mode)
     A = np.zeros((3, 3), dtype=np_type)
@@ -96,9 +100,10 @@ def test_laplace_bilinear_form_2d(mode, expected_result, compile_args):
          dtype=np.complex64)),
 ])
 def test_mass_bilinear_form_2d(mode, expected_result, compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     a = ufl.inner(u, v) * ufl.dx
     L = ufl.conj(v) * ufl.dx
     forms = [a, L]
@@ -108,8 +113,8 @@ def test_mass_bilinear_form_2d(mode, expected_result, compile_args):
     for f, compiled_f in zip(forms, compiled_forms):
         assert compiled_f.rank == len(f.arguments())
 
-    form0 = compiled_forms[0].integrals(module.lib.cell)[0]
-    form1 = compiled_forms[1].integrals(module.lib.cell)[0]
+    form0 = compiled_forms[0].form_integrals[0]
+    form1 = compiled_forms[1].form_integrals[0]
 
     np_type = cdtype_to_numpy(mode)
     A = np.zeros((3, 3), dtype=np_type)
@@ -149,9 +154,10 @@ def test_mass_bilinear_form_2d(mode, expected_result, compile_args):
      - (1.0j / 24.0) * np.array([[2, 1, 1], [1, 2, 1], [1, 1, 2]], dtype=np.complex128)),
 ])
 def test_helmholtz_form_2d(mode, expected_result, compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     if mode == "double":
         k = 1.0
     elif mode == "double _Complex":
@@ -167,7 +173,7 @@ def test_helmholtz_form_2d(mode, expected_result, compile_args):
     for f, compiled_f in zip(forms, compiled_forms):
         assert compiled_f.rank == len(f.arguments())
 
-    form0 = compiled_forms[0].integrals(module.lib.cell)[0]
+    form0 = compiled_forms[0].form_integrals[0]
 
     np_type = cdtype_to_numpy(mode)
     A = np.zeros((3, 3), dtype=np_type)
@@ -205,9 +211,10 @@ def test_helmholtz_form_2d(mode, expected_result, compile_args):
          dtype=np.complex128)),
 ])
 def test_laplace_bilinear_form_3d(mode, expected_result, compile_args):
-    cell = ufl.tetrahedron
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "tetrahedron", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "tetrahedron", 1, shape=(3, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     a = ufl.inner(ufl.grad(u), ufl.grad(v)) * ufl.dx
     forms = [a]
     compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(
@@ -216,7 +223,7 @@ def test_laplace_bilinear_form_3d(mode, expected_result, compile_args):
     for f, compiled_f in zip(forms, compiled_forms):
         assert compiled_f.rank == len(f.arguments())
 
-    form0 = compiled_forms[0].integrals(module.lib.cell)[0]
+    form0 = compiled_forms[0].form_integrals[0]
 
     np_type = cdtype_to_numpy(mode)
     A = np.zeros((4, 4), dtype=np_type)
@@ -242,10 +249,11 @@ def test_laplace_bilinear_form_3d(mode, expected_result, compile_args):
 
 
 def test_form_coefficient(compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TestFunction(element), ufl.TrialFunction(element)
-    g = ufl.Coefficient(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TestFunction(space), ufl.TrialFunction(space)
+    g = ufl.Coefficient(space)
     a = g * ufl.inner(u, v) * ufl.dx
     forms = [a]
     compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(forms, cffi_extra_compile_args=compile_args)
@@ -253,7 +261,7 @@ def test_form_coefficient(compile_args):
     for f, compiled_f in zip(forms, compiled_forms):
         assert compiled_f.rank == len(f.arguments())
 
-    form0 = compiled_forms[0].integrals(module.lib.cell)[0]
+    form0 = compiled_forms[0].form_integrals[0]
     A = np.zeros((3, 3), dtype=np.float64)
     w = np.array([1.0, 1.0, 1.0], dtype=np.float64)
     c = np.array([], dtype=np.float64)
@@ -278,9 +286,10 @@ def test_form_coefficient(compile_args):
 
 
 def test_subdomains(compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     a0 = ufl.inner(u, v) * ufl.dx + ufl.inner(u, v) * ufl.dx(2)
     a1 = ufl.inner(u, v) * ufl.dx(2) + ufl.inner(u, v) * ufl.dx
     a2 = ufl.inner(u, v) * ufl.dx(2) + ufl.inner(u, v) * ufl.dx(1)
@@ -293,29 +302,35 @@ def test_subdomains(compile_args):
         assert compiled_f.rank == len(f.arguments())
 
     form0 = compiled_forms[0]
-    ids = form0.integral_ids(module.lib.cell)
+    offsets = form0.form_integral_offsets
+    cell = module.lib.cell
+    ids = [form0.form_integral_ids[j] for j in range(offsets[cell], offsets[cell + 1])]
     assert ids[0] == -1 and ids[1] == 2
 
     form1 = compiled_forms[1]
-    ids = form1.integral_ids(module.lib.cell)
+    offsets = form1.form_integral_offsets
+    ids = [form1.form_integral_ids[j] for j in range(offsets[cell], offsets[cell + 1])]
     assert ids[0] == -1 and ids[1] == 2
 
     form2 = compiled_forms[2]
-    ids = form2.integral_ids(module.lib.cell)
+    offsets = form2.form_integral_offsets
+    ids = [form2.form_integral_ids[j] for j in range(offsets[cell], offsets[cell + 1])]
     assert ids[0] == 1 and ids[1] == 2
 
     form3 = compiled_forms[3]
-    assert form3.num_integrals(module.lib.cell) == 0
-
-    ids = form3.integral_ids(module.lib.exterior_facet)
+    offsets = form3.form_integral_offsets
+    assert offsets[cell + 1] - offsets[cell] == 0
+    exf = module.lib.exterior_facet
+    ids = [form3.form_integral_ids[j] for j in range(offsets[exf], offsets[exf + 1])]
     assert ids[0] == 0 and ids[1] == 210
 
 
 @pytest.mark.parametrize("mode", ["double", "double _Complex"])
 def test_interior_facet_integral(mode, compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     a0 = ufl.inner(ufl.jump(ufl.grad(u)), ufl.jump(ufl.grad(v))) * ufl.dS
     forms = [a0]
     compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(
@@ -331,7 +346,7 @@ def test_interior_facet_integral(mode, compile_args):
     ffi = module.ffi
     np_type = cdtype_to_numpy(mode)
 
-    integral0 = form0.integrals(module.lib.interior_facet)[0]
+    integral0 = form0.form_integrals[0]
     A = np.zeros((6, 6), dtype=np_type)
     w = np.array([], dtype=np_type)
     c = np.array([], dtype=np.float64)
@@ -359,10 +374,11 @@ def test_interior_facet_integral(mode, compile_args):
 
 @pytest.mark.parametrize("mode", ["double", "double _Complex"])
 def test_conditional(mode, compile_args):
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    u, v = ufl.TrialFunction(element), ufl.TestFunction(element)
-    x = ufl.SpatialCoordinate(cell)
+    element = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
+    x = ufl.SpatialCoordinate(domain)
     condition = ufl.Or(ufl.ge(ufl.real(x[0] + x[1]), 0.1),
                        ufl.ge(ufl.real(x[1] + x[1]**2), 0.1))
     c1 = ufl.conditional(condition, 2.0, 1.0)
@@ -377,8 +393,8 @@ def test_conditional(mode, compile_args):
     compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(
         forms, options={'scalar_type': mode}, cffi_extra_compile_args=compile_args)
 
-    form0 = compiled_forms[0].integrals(module.lib.cell)[0]
-    form1 = compiled_forms[1].integrals(module.lib.cell)[0]
+    form0 = compiled_forms[0].form_integrals[0]
+    form1 = compiled_forms[1].form_integrals[0]
 
     ffi = module.ffi
     np_type = cdtype_to_numpy(mode)
@@ -417,10 +433,10 @@ def test_conditional(mode, compile_args):
 
 
 def test_custom_quadrature(compile_args):
-    ve = ufl.VectorElement("P", "triangle", 1)
+    ve = basix.ufl.element("P", "triangle", 1, shape=(2, ))
     mesh = ufl.Mesh(ve)
 
-    e = ufl.FiniteElement("P", mesh.ufl_cell(), 2)
+    e = basix.ufl.element("P", mesh.ufl_cell().cellname(), 2)
     V = ufl.FunctionSpace(mesh, e)
     u, v = ufl.TrialFunction(V), ufl.TestFunction(V)
 
@@ -434,7 +450,7 @@ def test_custom_quadrature(compile_args):
 
     ffi = module.ffi
     form = compiled_forms[0]
-    default_integral = form.integrals(module.lib.cell)[0]
+    default_integral = form.form_integrals[0]
 
     A = np.zeros((6, 6), dtype=np.float64)
     w = np.array([], dtype=np.float64)
@@ -455,8 +471,10 @@ def test_custom_quadrature(compile_args):
 
 
 def test_curl_curl(compile_args):
-    V = ufl.FiniteElement("N1curl", "triangle", 2)
-    u, v = ufl.TrialFunction(V), ufl.TestFunction(V)
+    V = basix.ufl.element("N1curl", "triangle", 2)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, V)
+    u, v = ufl.TrialFunction(space), ufl.TestFunction(space)
     a = ufl.inner(ufl.curl(u), ufl.curl(v)) * ufl.dx
 
     forms = [a]
@@ -508,9 +526,10 @@ def lagrange_triangle_symbolic(order, corners=[(1, 0), (2, 0), (0, 1)], fun=lamb
 @pytest.mark.parametrize("order", [1, 2, 3])
 def test_lagrange_triangle(compile_args, order, mode, sym_fun, ufl_fun):
     sym = lagrange_triangle_symbolic(order, fun=sym_fun)
-    cell = ufl.triangle
-    element = ufl.FiniteElement("Lagrange", cell, order)
-    v = ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "triangle", order)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, element)
+    v = ufl.TestFunction(space)
 
     a = ufl_fun(v) * ufl.dx
     forms = [a]
@@ -520,20 +539,17 @@ def test_lagrange_triangle(compile_args, order, mode, sym_fun, ufl_fun):
     ffi = module.ffi
     form0 = compiled_forms[0]
 
-    assert form0.num_integrals(module.lib.cell) == 1
-    default_integral = form0.integrals(module.lib.cell)[0]
+    assert form0.form_integral_offsets[module.lib.cell + 1] == 1
+    default_integral = form0.form_integrals[0]
 
     np_type = cdtype_to_numpy(mode)
     b = np.zeros((order + 2) * (order + 1) // 2, dtype=np_type)
     w = np.array([], dtype=np_type)
-
     geom_type = scalar_to_value_type(mode)
     np_gtype = cdtype_to_numpy(geom_type)
-
     coords = np.array([[1.0, 0.0, 0.0],
                        [2.0, 0.0, 0.0],
                        [0.0, 1.0, 0.0]], dtype=np_gtype)
-
     kernel = getattr(default_integral, f"tabulate_tensor_{np_type}")
     kernel(ffi.cast('{type} *'.format(type=mode), b.ctypes.data),
            ffi.cast('{type} *'.format(type=mode), w.ctypes.data),
@@ -600,9 +616,10 @@ def lagrange_tetrahedron_symbolic(order, corners=[(1, 0, 0), (2, 0, 0), (0, 1, 0
 @pytest.mark.parametrize("order", [1, 2, 3])
 def test_lagrange_tetrahedron(compile_args, order, mode, sym_fun, ufl_fun):
     sym = lagrange_tetrahedron_symbolic(order, fun=sym_fun)
-    cell = ufl.tetrahedron
-    element = ufl.FiniteElement("Lagrange", cell, order)
-    v = ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "tetrahedron", order)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "tetrahedron", 1, shape=(3, )))
+    space = ufl.FunctionSpace(domain, element)
+    v = ufl.TestFunction(space)
 
     a = ufl_fun(v) * ufl.dx
     forms = [a]
@@ -612,9 +629,9 @@ def test_lagrange_tetrahedron(compile_args, order, mode, sym_fun, ufl_fun):
     ffi = module.ffi
     form0 = compiled_forms[0]
 
-    assert form0.num_integrals(module.lib.cell) == 1
+    assert form0.form_integral_offsets[module.lib.cell + 1] == 1
 
-    default_integral = form0.integrals(module.lib.cell)[0]
+    default_integral = form0.form_integrals[0]
 
     np_type = cdtype_to_numpy(mode)
     b = np.zeros((order + 3) * (order + 2) * (order + 1) // 6, dtype=np_type)
@@ -639,9 +656,10 @@ def test_lagrange_tetrahedron(compile_args, order, mode, sym_fun, ufl_fun):
 
 
 def test_prism(compile_args):
-    cell = ufl.prism
-    element = ufl.FiniteElement("Lagrange", cell, 1)
-    v = ufl.TestFunction(element)
+    element = basix.ufl.element("Lagrange", "prism", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "prism", 1, shape=(3, )))
+    space = ufl.FunctionSpace(domain, element)
+    v = ufl.TestFunction(space)
 
     L = v * ufl.dx
     forms = [L]
@@ -650,9 +668,9 @@ def test_prism(compile_args):
 
     ffi = module.ffi
     form0 = compiled_forms[0]
-    assert form0.num_integrals(module.lib.cell) == 1
+    assert form0.form_integral_offsets[module.lib.cell + 1] == 1
 
-    default_integral = form0.integrals(module.lib.cell)[0]
+    default_integral = form0.form_integrals[0]
     b = np.zeros(6, dtype=np.float64)
     coords = np.array([1.0, 0.0, 0.0,
                        0.0, 1.0, 0.0,
@@ -672,10 +690,10 @@ def test_prism(compile_args):
 
 def test_complex_operations(compile_args):
     mode = "double _Complex"
-    cell = ufl.triangle
-    c_element = ufl.VectorElement("Lagrange", cell, 1)
+    cell = "triangle"
+    c_element = basix.ufl.element("Lagrange", cell, 1, shape=(2, ))
     mesh = ufl.Mesh(c_element)
-    element = ufl.VectorElement("DG", cell, 0)
+    element = basix.ufl.element("DG", cell, 0, shape=(2, ))
     V = ufl.FunctionSpace(mesh, element)
     u = ufl.Coefficient(V)
     J1 = ufl.real(u)[0] * ufl.imag(u)[1] * ufl.conj(u)[0] * ufl.dx
@@ -685,8 +703,8 @@ def test_complex_operations(compile_args):
     compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(
         forms, options={'scalar_type': mode}, cffi_extra_compile_args=compile_args)
 
-    form0 = compiled_forms[0].integrals(module.lib.cell)[0]
-    form1 = compiled_forms[1].integrals(module.lib.cell)[0]
+    form0 = compiled_forms[0].form_integrals[0]
+    form1 = compiled_forms[1].form_integrals[0]
 
     ffi = module.ffi
     np_type = cdtype_to_numpy(mode)
@@ -728,8 +746,10 @@ def test_invalid_function_name(compile_args):
     old_str = ufl.Coefficient.__str__
     ufl.Coefficient.__str__ = lambda self: "invalid function name"
 
-    V = ufl.FiniteElement("Lagrange", ufl.triangle, 1)
-    u = ufl.Coefficient(V)
+    V = basix.ufl.element("Lagrange", "triangle", 1)
+    domain = ufl.Mesh(basix.ufl.element("Lagrange", "triangle", 1, shape=(2, )))
+    space = ufl.FunctionSpace(domain, V)
+    u = ufl.Coefficient(space)
     a = ufl.inner(u, u) * ufl.dx
 
     forms = [a]
@@ -744,3 +764,162 @@ def test_invalid_function_name(compile_args):
 
     # Revert monkey patch for other tests
     ufl.Coefficient.__str__ = old_str
+
+
+def test_interval_vertex_quadrature(compile_args):
+
+    c_el = basix.ufl.element("Lagrange", "interval", 1, shape=(1, ))
+    mesh = ufl.Mesh(c_el)
+
+    x = ufl.SpatialCoordinate(mesh)
+    dx = ufl.Measure(
+        "dx", metadata={"quadrature_rule": "vertex"})
+    b = x[0] * dx
+
+    forms = [b]
+    compiled_forms, module, code = ffcx.codegeneration.jit.compile_forms(
+        forms, cffi_extra_compile_args=compile_args)
+
+    ffi = module.ffi
+    form0 = compiled_forms[0]
+    assert form0.form_integral_offsets[module.lib.cell + 1] == 1
+
+    default_integral = form0.form_integrals[0]
+    J = np.zeros(1, dtype=np.float64)
+    a = np.pi
+    b = np.exp(1)
+    coords = np.array([a, 0.0, 0.0,
+
+                       b, 0.0, 0.0], dtype=np.float64)
+
+    kernel = getattr(default_integral, "tabulate_tensor_float64")
+    kernel(ffi.cast('double *', J.ctypes.data),
+           ffi.NULL,
+           ffi.NULL,
+           ffi.cast('double *', coords.ctypes.data), ffi.NULL, ffi.NULL)
+    assert np.isclose(J[0], (0.5 * a + 0.5 * b) * np.abs(b - a))
+
+
+def test_facet_vertex_quadrature(compile_args):
+    """Test facet vertex quadrature"""
+    c_el = basix.ufl.element("Lagrange", "quadrilateral", 1, shape=(2,))
+    mesh = ufl.Mesh(c_el)
+
+    x = ufl.SpatialCoordinate(mesh)
+    ds = ufl.Measure(
+        "ds", metadata={"quadrature_rule": "vertex"})
+    expr = (x[0] + ufl.cos(x[1]))
+    b1 = expr * ds
+    ds_c = ufl.Measure(
+        "ds",
+        metadata={
+            "quadrature_rule": "custom",
+            "quadrature_points": np.array([[0.0], [1.0]]),
+            "quadrature_weights": np.array([1.0 / 2.0, 1.0 / 2.0]),
+        }
+    )
+    b2 = expr * ds_c
+    forms = [b1, b2]
+    compiled_forms, module, _ = ffcx.codegeneration.jit.compile_forms(
+        forms, cffi_extra_compile_args=compile_args)
+
+    ffi = module.ffi
+    assert len(compiled_forms) == 2
+    solutions = []
+    for form in compiled_forms:
+        offsets = form.form_integral_offsets
+        exf = module.lib.exterior_facet
+        assert offsets[exf + 1] - offsets[exf] == 1
+
+        default_integral = form.form_integrals[offsets[exf]]
+        J = np.zeros(1, dtype=np.float64)
+        a = np.pi
+        b = np.exp(1)
+        coords = np.array([a, 0.1, 0.0,
+                           a + b, 0.0, 0.0,
+                           a, a, 0.,
+                           a + 2 * b, a, 0.], dtype=np.float64)
+        # First facet is between vertex 0 and 1 in coords
+        facets = np.array([0], dtype=np.intc)
+
+        kernel = getattr(default_integral, "tabulate_tensor_float64")
+        kernel(ffi.cast('double *', J.ctypes.data),
+               ffi.NULL,
+               ffi.NULL,
+               ffi.cast('double *', coords.ctypes.data),
+               ffi.cast('int *', facets.ctypes.data),
+               ffi.NULL)
+        solutions.append(J[0])
+        # Test against exact result
+        assert np.isclose(J[0], (0.5 * (a + np.cos(0.1)) + 0.5 * (a + b + np.cos(0))) * np.sqrt(b**2 + 0.1**2))
+
+    # Compare custom quadrature with vertex quadrature
+    assert np.isclose(solutions[0], solutions[1])
+
+
+def test_manifold_derivatives(compile_args):
+    """Test higher order derivatives on manifolds"""
+    c_el = basix.ufl.element("Lagrange", "interval", 1, shape=(2,), gdim=2)
+    mesh = ufl.Mesh(c_el)
+
+    x = ufl.SpatialCoordinate(mesh)
+    dx = ufl.Measure("dx", domain=mesh)
+    order = 4
+    el = basix.ufl.element("Lagrange", "interval", order, gdim=2)
+    V = ufl.FunctionSpace(mesh, el)
+
+    u = ufl.Coefficient(V)
+    d = 5.3
+    f_ex = d * order * (order - 1) * x[1]**(order - 2)
+    expr = u.dx(1).dx(1) - f_ex
+    J = expr * expr * dx
+
+    compiled_forms, module, _ = ffcx.codegeneration.jit.compile_forms(
+        [J], cffi_extra_compile_args=compile_args)
+
+    default_integral = compiled_forms[0].form_integrals[0]
+    scale = 2.5
+    coords = np.array([0.0, 0.0, 0.0, 0.0, scale, 0.0], dtype=np.float64)
+    dof_coords = scale * el.element.points.reshape(-1)
+
+    w = np.array([d * d_c**order for d_c in dof_coords], dtype=np.float64)
+    c = np.array([], dtype=np.float64)
+    perm = np.array([0], dtype=np.uint8)
+
+    ffi = module.ffi
+    J = np.zeros(1, dtype=np.float64)
+    kernel = getattr(default_integral, "tabulate_tensor_float64")
+    kernel(ffi.cast('double *', J.ctypes.data),
+           ffi.cast('double  *', w.ctypes.data),
+           ffi.cast('double  *', c.ctypes.data),
+           ffi.cast('double  *', coords.ctypes.data), ffi.NULL,
+           ffi.cast('uint8_t *', perm.ctypes.data))
+
+    assert np.isclose(J[0], 0.0)
+
+
+def test_integral_grouping(compile_args):
+    """We group integrals with common integrands to avoid duplicated
+    integration kernels. This means that `inner(u, v)*dx((1,2,3))  +
+    inner(grad(u), grad(v))*dx(2) + inner(u,v)*dx` is grouped as
+    1. `inner(u,v)*dx(("everywhere", 1, 3))`
+    2. `(inner(grad(u), grad(v)) + inner(u, v))*dx(2)`
+    Each of the forms has one generated `tabulate_tensor_*` function,
+    which is referred to multiple times in `integrals_` and
+    `integral_ids_`
+
+    """
+    mesh = ufl.Mesh(ufl.VectorElement("Lagrange", ufl.triangle, 1))
+    V = ufl.FunctionSpace(mesh, ufl.FiniteElement("Lagrange", ufl.triangle, 1))
+    u = ufl.TrialFunction(V)
+    v = ufl.TestFunction(V)
+    a = ufl.inner(u, v) * ufl.dx((1, 2, 3)) + ufl.inner(ufl.grad(u), ufl.grad(v)) * ufl.dx(2) + ufl.inner(u, v) * ufl.dx
+    compiled_forms, module, _ = ffcx.codegeneration.jit.compile_forms(
+        [a], cffi_extra_compile_args=compile_args)
+    # NOTE: This assumes that the first integral type is cell integrals, see UFCx.h
+    cell = module.lib.cell
+    num_integrals = compiled_forms[0].form_integral_offsets[cell + 1] - compiled_forms[0].form_integral_offsets[cell]
+    assert num_integrals == 4
+    unique_integrals = set([compiled_forms[0].form_integrals[compiled_forms[0].form_integral_offsets[cell] + i]
+                            for i in range(num_integrals)])
+    assert len(unique_integrals) == 2
